==> 04_embedding_vis1_PCA.Rmd <==
---
title: "embedding visualization with PCA"
output: html_document
---


```{r setup, include=FALSE}
library(tidyverse)
library(foreign)
library(haven)
library(readr)
library(Rtsne)
library(ggplot2)
library(corrplot)
library(factoextra)
```

# Data import 
Continued from embedding_evaluating.Rmd 

## 1. Import more covariates 
```{r}
# importing more covariates from 2003-2004
bp_d1 = read_xpt("001_raw_data/BPX_C.xpt") |>
  janitor::clean_names() |>
  rowwise() |>
  mutate(
    med_sbp = median(c(bpxsy1, bpxsy2, bpxsy3, bpxsy4), na.rm = TRUE),
    med_dbp = median(c(bpxdi1, bpxdi2, bpxdi3, bpxdi4), na.rm = TRUE)
  ) |>
  select(seqn, med_sbp, med_dbp) |> # blood pressure in mmHg
  drop_na()

### check if it is okay to use average BP without considering the enhancement usage after each reading? 

ldl_d1 = read_xpt("001_raw_data/L13AM_C.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbdldl)

choles_d1 = read_xpt("001_raw_data/L13_C.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbxtc, lbxhdd) |>
  left_join(ldl_d1, by = "seqn") |>
  rename(chol_total = lbxtc, 
         chol_hdl = lbxhdd, 
         chol_ldl = lbdldl) # mg/dL

diabetes_d1 = read_xpt("001_raw_data/DIQ_C.xpt") |>
  janitor::clean_names() |>
  select(seqn, diq010) |>
  filter(diq010 !=7 & diq010 !=9) |> 
  rename(diabetes = diq010) # 1=yes, 2=no, 3=borderline, 7=refused, 9=dontknow

mc_d1 = read_xpt("001_raw_data/MCQ_C.xpt") |>
  janitor::clean_names() |>
  select(seqn, mcq160a, mcq160j, mcq220) |>
  filter(mcq160a != 7 & mcq160a != 9 & mcq160j != 7 & mcq160j != 9 & 
           mcq220 != 7 & mcq220 != 9) |> 
  rename(arthritis = mcq160a, 
         overweight = mcq160j, 
         malig = mcq220) # 1=yes, 2=no, 7=refused, 9=dontknow 

cov_d1 = bp_d1 |> 
  left_join(choles_d1, by = "seqn") |>
  left_join(diabetes_d1, by = "seqn") |>
  left_join(mc_d1, by ="seqn")


# importing more covariates from 2005-2006
bp_d2 = read_xpt("001_raw_data/BPX_D.xpt") |>
  janitor::clean_names() |>
  rowwise() |>
  mutate(
    med_sbp = median(c(bpxsy1, bpxsy2, bpxsy3, bpxsy4), na.rm = TRUE),
    med_dbp = median(c(bpxdi1, bpxdi2, bpxdi3, bpxdi4), na.rm = TRUE)
  ) |>
  select(seqn, med_sbp, med_dbp) |> #blood pressure in mmHg
  drop_na()

ldl_d2 = read_xpt("001_raw_data/TRIGLY_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbdldl)

hdl_d2 = read_xpt("001_raw_data/HDL_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbdhdd)

choles_d2 = read_xpt("001_raw_data/TCHOL_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbxtc) |>
  left_join(ldl_d2, by = "seqn") |>
  left_join(hdl_d2, by = "seqn") |>
  rename(chol_total = lbxtc, 
         chol_hdl = lbdhdd, 
         chol_ldl = lbdldl) # mg/dL

diabetes_d2 = read_xpt("001_raw_data/DIQ_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, diq010) |>
  filter(diq010 !=7 & diq010 !=9) |> 
  rename(diabetes = diq010) # 1=yes, 2=no, 3=borderline, 7=refused, 9=dontknow

mc_d2 = read_xpt("001_raw_data/MCQ_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, mcq160a, mcq220, mcq080) |>
  filter(mcq160a != 7 & mcq160a != 9 & mcq220 != 7 & mcq220 != 9, mcq080!=7 & mcq080!=9) |> 
  rename(arthritis = mcq160a, 
         overweight = mcq080, 
         malig = mcq220) # 1=yes, 2=no, 7=refused, 9=dontknow 

cov_d2 = bp_d2 |> 
  left_join(choles_d2, by = "seqn") |>
  left_join(diabetes_d2, by ="seqn") |>
  left_join(mc_d2, by = "seqn")



# clean up covariates 
covariates = rbind(cov_d1, cov_d2) |>
  mutate(
    sbp_cat = case_when(med_sbp < 130 ~ 0, 
                          med_sbp >= 130 & med_sbp < 140 ~ 1, 
                          med_sbp >= 140 & med_sbp < 180~ 2, 
                          med_sbp >= 180 ~ 3), 
    dbp_cat = case_when(med_dbp < 80 ~ 0, 
                          med_dbp >= 80 & med_dbp < 90 ~ 1, 
                          med_dbp >= 90 & med_dbp < 120 ~ 2, 
                          med_dbp <= 120 ~ 3), 
    chol_total_cat = case_when(chol_total < 200 ~ 0, 
                                 chol_total >= 200 & chol_total< 240 ~ 1, 
                                 chol_total >= 240 ~ 2), 
    chol_ldl_cat = case_when(chol_ldl < 130 ~ 0, 
                               chol_ldl >= 130 & chol_ldl< 160 ~ 1, 
                               chol_ldl >= 160 ~ 2), 
    chol_hdl_cat = case_when(chol_hdl >= 60 ~ 0, 
                               chol_hdl >= 50 & chol_hdl< 60 ~ 1, 
                               chol_hdl < 50 ~ 2))
```

Variables: 

1) med_sbp: median systolic blood pressure in mmHg
2) med_dbp: median diastolic blood pressure in mmHg
3) chol_total: choleterol in mg/dL
4) chol_ldl: LDL in mg/dL
5) chol_hdl: HDL in mg/dL 
6) diabetes: ever told to have diabetes by doctor (1=yes, 2=no, 3=borderline)
7) arthritis: ever told to have arthritis by doctor (1=yes, 2=no)
8) overweight: ever told to be overweight by doctor (1=yes, 2=no)
9) malig: ever told to have cancer of malignancy by doctor (1=yes, 2=no) 

10) sbp_cat: 0=low risk, 1=stage1 hypertension, 2=stage2 hypertension, 3=crisis
11) dbp_cat: 0=low risk, 1=stage1 hypertension, 2=stage2 hypertension, 3=crisis
12) chol_total_cat: 0=low risk, 1=borderline, 2=high risk
13) chol_ldl_cat: 0=low risk, 1=borderline, 2=high risk
14) chol_hdl_cat: 0=low risk, 1=borderline, 2=high risk


## 2. Import embeddings of other models 
```{r embedding_import}
# GPT with 1536 embedding dimension 
data_gpt1536 = read.csv("./003_entro_data/data_wide_embedding_gpt1536.csv") |>
  janitor::clean_names() |>
  dplyr::select(-x,-combined)

data_gpt1536$embedding = gsub("\\[", "", data_gpt1536$embedding)
data_gpt1536$embedding = gsub("\\]", "", data_gpt1536$embedding)

data_gpt1536 = data_gpt1536 |>
  separate(embedding, into = paste0("var", 1:1536), sep = ",\\s*", convert = TRUE) |>
  mutate(across(starts_with("var"), as.numeric))


# BERT with 768 embedding dimension
data_bert768 = read.csv("./003_entro_data/data_wide_embedding_bert768.csv") |>
  janitor::clean_names() |>
  dplyr::select(-x,-combined)

data_bert768$embedding <- gsub("\\[", "", data_bert768$embedding)
data_bert768$embedding <- gsub("\\]", "", data_bert768$embedding)
data_bert768$embedding <- gsub("\n", " ", data_bert768$embedding)

data_bert768 = data_bert768 |>
  mutate(embedding = str_trim(embedding),  # Remove leading and trailing spaces
         embedding = str_replace_all(embedding, "\\s+", " ")) |> # Replace multiple spaces with a single space 
  separate(embedding, into = paste0("var", 1:768), sep = "\\s+", convert = TRUE) |>
  mutate(across(starts_with("var"), as.numeric))


# Cohere with 1024 embedding dimension
data_cohere1024 = read.csv("./003_entro_data/data_wide_embedding_cohere1024.csv") |>
  janitor::clean_names() |>
  dplyr::select(-x,-n_tokens)

data_cohere1024$embedding = gsub("\\[", "", data_cohere1024$embedding) 
data_cohere1024$embedding = gsub("\\]", "", data_cohere1024$embedding)
data_cohere1024$embedding = gsub(",", " ", data_cohere1024$embedding)

data_cohere1024 = data_cohere1024 |> 
  mutate(embedding = str_trim(embedding),  # Remove leading and trailing spaces
         embedding = str_replace_all(embedding, "\\s+", " ")) |> # Replace multiple spaces with a single space
  separate(embedding, into = paste0("var", 1:1024), sep = "\\s+", convert = TRUE) |>
  mutate(across(starts_with("var"), as.numeric))
```

## 3. Import the entropy 
```{r}
# Entropy
data_entropy = read.csv("./003_entro_data/data_wide_entropy.csv") |>
  janitor::clean_names() |>
  dplyr::select(-x) 

# GPT1536 + entropy
data_gpt1536_entropy = data_entropy |>
  dplyr::select(seqn, entropy_day1:entropy_day7) |> 
  inner_join(data_gpt1536, by = "seqn") 

# BERT768 + entropy
data_bert768_entropy = data_entropy |>
  dplyr::select(seqn, entropy_day1:entropy_day7) |> 
  inner_join(data_bert768, by = "seqn") 

# Cohere1024 + entropy
data_cohere1024_entropy = data_entropy |>
  dplyr::select(seqn, entropy_day1:entropy_day7) |> 
  inner_join(data_cohere1024, by = "seqn") 
```

## 4. Get MOMENT embeddings  
```{r}
# MOMENT with 1024 embedding dimension  
data1_moment1024 = read.csv("./002_data/embeddings_moment_subset1_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn) 

data2_moment1024 = read.csv("./002_data/embeddings_moment_subset2_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn) 

data_moment1024 = rbind(data1_moment1024, data2_moment1024) 
```

## 5. Get MOMENT RAW DATA embeddings 
```{r}
# MOMENT raw data embedding with 1024 dimension  
data1_raw_moment1024 = read.csv("./002_data/embeddings_moment_raw_subset1_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn)

data2_raw_moment1024 = read.csv("./002_data/embeddings_moment_raw_subset2_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn) 

data_raw_moment1024 = rbind(data1_raw_moment1024, data2_raw_moment1024) 
```

## 6. Get MOMENT RAW RECODED embeddings 
```{r}
# MOMENT raw recoded embedding with 1024 dimension  
data1_recode_moment1024 = read.csv("./002_data/embeddings_moment_recoded_subset1_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn)

data2_recode_moment1024 = read.csv("./002_data/embeddings_moment_recoded_subset2_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn) 

data_recode_moment1024 = rbind(data1_recode_moment1024, data2_recode_moment1024) 
```




# Visualization - PCA with only demographics and scaled emebddings  

## 1. PCA of EntroGPT 
### 1.1 PCA of EntroGPT : color coded by response or demographics 
```{r PCA_EntroGPT}
################ Function for PCA for this data with demographics 
pca_analysis = function(data, embed_cols, demo_cols, response_col, id_col = "seqn", color_mapping = list()) {
  
  # Function to get appropriate labels based on the column
  get_labels = function(col_name, values, color_mapping) {
    if (is.null(color_mapping[[col_name]])) {
      
      if (col_name == "bmi") {
        return(ifelse(values == 1, "BMI>25", "Otherwise"))
      } else if (col_name == "gender") {
        return(ifelse(values == 1, "male", "female"))
      } else if (col_name == "race") {
        return(factor(values, levels = 1:5,
                     labels = c("Mexican American", "Other Hispanic", 
                               "Non-Hispanic White", 
                               "Non-Hispanic Black", "Other")))
      } else if (col_name == "education") {
        return(factor(values, levels = 1:5,
                     labels = c("less than 9th grade", "9-11th grade", 
                               "high school grad/GED or equivalent", 
                               "some college or associates degree", 
                               "college graduate or higher")))
      } else if (col_name == "married") {
        return(factor(values, levels = 1:6,
                     labels = c("married", "widowed", "divorced", 
                                "separated", "never married", 
                                "living with partner")))
      } else if (col_name == "age_cat") {
        return(factor(values, levels = 0:2,
                     labels = c("young", "middle", "older")))
      } else {
        return(as.factor(values))
      }
      
    } else {
      mapping = color_mapping[[col_name]]
      if (is.function(mapping)) {
        return(mapping(values))
      } else {
        return(mapping[values])
      }
    }
  }
  
  # Separate into embeddings and covariates/response
  embed_data = data[, c(embed_cols)]
  covar_data = data[, c(id_col, demo_cols, response_col)]
  if ("age_cat" %in% names(data) && !("age_cat" %in% demo_cols)) {
    covar_data$age_cat = data$age_cat
  }
  
  # Scale the embeddings
  embed_data_scaled = as.data.frame(scale(embed_data))
  embed_data_scaled[[id_col]] = data[[id_col]]
  
  # Combine scaled embeddings with covariates
  data_scaled = embed_data_scaled |>
    left_join(covar_data, by = id_col)
  
  # Create PCA model  
  pca_vars = data_scaled |>
    select(all_of(c(embed_cols)))
  
  pca_model = prcomp(pca_vars, scale. = TRUE)
  
  # Create PCAs 
  pca_results = list()
  
  # PCA plot for response variable bmi
  pca_results[[response_col]] = list(
    model = pca_model,
    plot = fviz_pca_ind(
      pca_model,
      habillage = get_labels(response_col, data_scaled[[response_col]], color_mapping),
      label = "var",
      title = paste("PCA colored by", response_col)
    )
  )
  
  # PCA plots for each demographic variable
  for (demo_col in demo_cols) {
    # special case for age 
    color_col = demo_col
    if (demo_col == "age" && "age_cat" %in% names(data_scaled)) {
      color_col <- "age_cat"
    }
    
    pca_results[[demo_col]] = list(
      model = pca_model,   
      plot = fviz_pca_ind(
        pca_model,
        habillage = get_labels(color_col, data_scaled[[color_col]], color_mapping),
        label = "var",
        title = paste("PCA colored by", ifelse(color_col == "age_cat" && demo_col == "age", 
                                             "age (categorized)", color_col))
      )
    )
  }
  
  # Return results
  return(list(
    data_scaled = data_scaled,
    pca_model = pca_model,
    pca_results = pca_results
  ))
}


# change continuous age to categories 
data_gpt1536_entropy = data_gpt1536_entropy |>
    mutate(age_cat = case_when(age >= 20 & age <= 39 ~ 0,
                               age >= 40 & age <= 64 ~ 1,
                               age >= 65             ~ 2))

# define column groups for EntroGPT 
entrogpt_embed_cols = c(2:8, 16:1551)
entrogpt_demo_cols = c("gender", "race", "education", "married", "age") 
entrogpt_response_col = "bmi" # Response variable
entrogpt_id_col = "seqn" # ID column
  
# PCA for EntroGPT 
entrogpt_pca_results = pca_analysis(
  data = data_gpt1536_entropy,
  embed_cols = names(data_gpt1536_entropy)[entrogpt_embed_cols],
  demo_cols = entrogpt_demo_cols,
  response_col = entrogpt_response_col,
  id_col = entrogpt_id_col
)

entrogpt_pca_results$pca_results$bmi$plot
entrogpt_pca_results$pca_results$gender$plot
entrogpt_pca_results$pca_results$race$plot
entrogpt_pca_results$pca_results$education$plot
entrogpt_pca_results$pca_results$married$plot
entrogpt_pca_results$pca_results$age$plot
```

### 1.2 PCA of EntroGPT : color coded by additional covarates 
```{r PCA_EntroGPT color coded}
entrogpt_sub_embed = data_gpt1536_entropy |>
  select(c(1:8, 16:1551)) |>
  mutate(across(-seqn, ~ as.numeric(scale(.x))))

entrogpt_sub_cov = data_gpt1536_entropy[c(1,9:15, 1552)] |>
  left_join(covariates, by = "seqn")  

entrogpt_scaled_cov = entrogpt_sub_embed |>
  left_join(entrogpt_sub_cov, by = "seqn") |>
  drop_na()


# PCA with embeddings/demographics/bmi of 2914 data points 
pca_entrogpt_vars_cov = entrogpt_scaled_cov |>
  select(c(2:1551)) #1:seqn, 1552:age_cat 
pca_entrogpt_cov = prcomp(pca_entrogpt_vars_cov, scale. = TRUE)


# color by systolic BP
fviz_pca_ind(pca_entrogpt_cov, 
             habillage = factor(entrogpt_scaled_cov$sbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by diastolic BP
fviz_pca_ind(pca_entrogpt_cov, 
             habillage = factor(entrogpt_scaled_cov$dbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by total cholesterol
fviz_pca_ind(pca_entrogpt_cov, 
             habillage = factor(entrogpt_scaled_cov$chol_total_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by HDL 
fviz_pca_ind(pca_entrogpt_cov, 
             habillage = factor(entrogpt_scaled_cov$chol_hdl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by LDL 
fviz_pca_ind(pca_entrogpt_cov, 
             habillage = factor(entrogpt_scaled_cov$chol_ldl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by diabetes  
fviz_pca_ind(pca_entrogpt_cov, 
             habillage = factor(entrogpt_scaled_cov$diabetes,
                               levels = 1:3,
                               labels = c("yes", "no", 
                                           "borderline")),  
             label = c("var"))

# color by arthritis  
fviz_pca_ind(pca_entrogpt_cov, 
             habillage = ifelse(entrogpt_scaled_cov$arthritis==1, 
                                "Arthritis", "No"), 
             label = c("var"))

# color by overweight  
fviz_pca_ind(pca_entrogpt_cov, 
             habillage = ifelse(entrogpt_scaled_cov$overweight==1, 
                                "Overweight", "No"), 
             label = c("var"))

# color by malignancy  
fviz_pca_ind(pca_entrogpt_cov, 
             habillage = ifelse(entrogpt_scaled_cov$malig==1, 
                                "Malignancy", "No"), 
             label = c("var"))




######### make function 

```


## 2. PCA of EntroBERT 
### 2.1 PCA of EntroBERT : color coded by response or demographics 
```{r PCA_EntroBERT}
# change continuous age to categories 
data_bert768_entropy = data_bert768_entropy |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# define column groups for EntroBERT 
entrobert_embed_cols = c(2:8, 16:783)
entrobert_demo_cols = c("gender", "race", "education", "married", "age") 
entrobert_response_col = "bmi" # Response variable
entrobert_id_col = "seqn"  
  

# PCA for EntroBERT 
entrobert_pca_results = pca_analysis(
  data = data_bert768_entropy,
  embed_cols = names(data_bert768_entropy)[entrobert_embed_cols],
  demo_cols = entrobert_demo_cols,
  response_col = entrobert_response_col,
  id_col = entrobert_id_col
)


entrobert_pca_results$pca_results$bmi$plot
entrobert_pca_results$pca_results$gender$plot
entrobert_pca_results$pca_results$race$plot
entrobert_pca_results$pca_results$education$plot
entrobert_pca_results$pca_results$married$plot
entrobert_pca_results$pca_results$age$plot
```

### 2.2 PCA of EntroBERT : color coded by additional covarates  
```{r PCA_EntroBERT color coded }
entrobert_sub_embed = data_bert768_entropy |>
  select(c(1:8, 16:783)) |>
  mutate(across(-seqn, ~ as.numeric(scale(.x))))

entrobert_sub_cov = data_bert768_entropy[c(1,9:15, 784)] |>
  left_join(covariates, by = "seqn")  

entrobert_scaled_cov = entrobert_sub_embed |>
  left_join(entrobert_sub_cov, by = "seqn") |>
  drop_na()


# PCA with embeddings/demographics/bmi of 2914 data points 
pca_entrobert_vars_cov = entrobert_scaled_cov |>
  select(c(2:783)) #1:seqn, 784:age_cat 
pca_entrobert_cov = prcomp(pca_entrobert_vars_cov, scale. = TRUE)


# color by systolic BP
fviz_pca_ind(pca_entrobert_cov, 
             habillage = factor(entrobert_scaled_cov$sbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by diastolic BP
fviz_pca_ind(pca_entrobert_cov, 
             habillage = factor(entrobert_scaled_cov$dbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by total cholesterol
fviz_pca_ind(pca_entrobert_cov, 
             habillage = factor(entrobert_scaled_cov$chol_total_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by HDL 
fviz_pca_ind(pca_entrobert_cov, 
             habillage = factor(entrobert_scaled_cov$chol_hdl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by LDL 
fviz_pca_ind(pca_entrobert_cov, 
             habillage = factor(entrobert_scaled_cov$chol_ldl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by diabetes  
fviz_pca_ind(pca_entrobert_cov, 
             habillage = factor(entrobert_scaled_cov$diabetes,
                               levels = 1:3,
                               labels = c("yes", "no", 
                                           "borderline")),  
             label = c("var"))

# color by arthritis  
fviz_pca_ind(pca_entrobert_cov, 
             habillage = ifelse(entrobert_scaled_cov$arthritis==1, 
                                "Arthritis", "No"), 
             label = c("var"))

# color by overweight  
fviz_pca_ind(pca_entrobert_cov, 
             habillage = ifelse(entrobert_scaled_cov$overweight==1, 
                                "Overweight", "No"), 
             label = c("var"))

# color by malignancy  
fviz_pca_ind(pca_entrobert_cov, 
             habillage = ifelse(entrobert_scaled_cov$malig==1, 
                                "Malignancy", "No"), 
             label = c("var"))
```


## 3. PCA of EntroCohere 
### 3.1 PCA of EntroCohere : color coded by response or demographics 
```{r PCA_EntroCohere}
# change continuous age to categories 
data_cohere1024_entropy = data_cohere1024_entropy |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# define column groups for EntroCohere 
entrocohere_embed_cols = c(2:8, 16:1039)
entrocohere_demo_cols = c("gender", "race", "education", "married", "age") 
entrocohere_response_col = "bmi" # Response variable
entrocohere_id_col = "seqn"  
  

# PCA for EntroCohere 
entrocohere_pca_results = pca_analysis(
  data = data_cohere1024_entropy,
  embed_cols = names(data_cohere1024_entropy)[entrocohere_embed_cols],
  demo_cols = entrocohere_demo_cols,
  response_col = entrocohere_response_col,
  id_col = entrocohere_id_col
)

entrocohere_pca_results$pca_results$bmi$plot
entrocohere_pca_results$pca_results$gender$plot
entrocohere_pca_results$pca_results$race$plot
entrocohere_pca_results$pca_results$education$plot
entrocohere_pca_results$pca_results$married$plot
entrocohere_pca_results$pca_results$age$plot
```

### 3.2 PCA of EntroCohere : color coded by additional covarates 
```{r PCA_EntroCohere color coded }
entrocohere_sub_embed = data_cohere1024_entropy |>
  select(c(1:8, 16:1039)) |>
  mutate(across(-seqn, ~ as.numeric(scale(.x))))

entrocohere_sub_cov = data_cohere1024_entropy[c(1,9:15, 1040)] |>
  left_join(covariates, by = "seqn")  

entrocohere_scaled_cov = entrocohere_sub_embed |>
  left_join(entrocohere_sub_cov, by = "seqn") |>
  drop_na()


# PCA with embeddings/demographics/bmi of 2914 data points 
pca_entrocohere_vars_cov = entrocohere_scaled_cov |>
  select(c(2:1039)) #1:seqn, 1040:age_cat 
pca_entrocohere_cov = prcomp(pca_entrocohere_vars_cov, scale. = TRUE)


# color by systolic BP
fviz_pca_ind(pca_entrocohere_cov, 
             habillage = factor(entrocohere_scaled_cov$sbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by diastolic BP
fviz_pca_ind(pca_entrocohere_cov, 
             habillage = factor(entrocohere_scaled_cov$dbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by total cholesterol
fviz_pca_ind(pca_entrocohere_cov, 
             habillage = factor(entrocohere_scaled_cov$chol_total_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by HDL 
fviz_pca_ind(pca_entrocohere_cov, 
             habillage = factor(entrocohere_scaled_cov$chol_hdl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by LDL 
fviz_pca_ind(pca_entrocohere_cov, 
             habillage = factor(entrocohere_scaled_cov$chol_ldl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by diabetes  
fviz_pca_ind(pca_entrocohere_cov, 
             habillage = factor(entrocohere_scaled_cov$diabetes,
                               levels = 1:3,
                               labels = c("yes", "no", 
                                           "borderline")),  
             label = c("var"))

# color by arthritis  
fviz_pca_ind(pca_entrocohere_cov, 
             habillage = ifelse(entrocohere_scaled_cov$arthritis==1, 
                                "Arthritis", "No"), 
             label = c("var"))

# color by overweight  
fviz_pca_ind(pca_entrocohere_cov, 
             habillage = ifelse(entrocohere_scaled_cov$overweight==1, 
                                "Overweight", "No"), 
             label = c("var"))

# color by malignancy  
fviz_pca_ind(pca_entrocohere_cov, 
             habillage = ifelse(entrocohere_scaled_cov$malig==1, 
                                "Malignancy", "No"), 
             label = c("var"))
```



## 4. PCA of MOMENT 
### 4.1 PCA of MOMENT : color coded by response or demographics 
```{r PCA_MOMENT}
# change continuous age to categories 
data_moment1024 = data_moment1024 |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# define column groups for MOMENT 
moment_embed_cols = c(9:1032)
moment_demo_cols = c("gender", "race", "education", "married", "age") 
moment_response_col = "bmi" # Response variable
moment_id_col = "seqn"  
  

# PCA for MOMENT 
moment_pca_results = pca_analysis(
  data = data_moment1024,
  embed_cols = names(data_moment1024)[moment_embed_cols],
  demo_cols = moment_demo_cols,
  response_col = moment_response_col,
  id_col = moment_id_col
)

moment_pca_results$pca_results$bmi$plot
moment_pca_results$pca_results$gender$plot
moment_pca_results$pca_results$race$plot
moment_pca_results$pca_results$education$plot
moment_pca_results$pca_results$married$plot
moment_pca_results$pca_results$age$plot
```

### 4.2 PCA of MOMENT : color coded by additional covarates 
```{r PCA_MOMENT color coded }
moment_sub_embed = data_moment1024 |>
  select(c(1, 9:1032)) |>
  mutate(across(-seqn, ~ as.numeric(scale(.x))))

moment_sub_cov = data_moment1024[c(1:8, 1033)] |>
  left_join(covariates, by = "seqn")  

moment_scaled_cov = moment_sub_embed |>
  left_join(moment_sub_cov, by = "seqn") |>
  drop_na()


# PCA with embeddings/demographics/bmi of 2914 data points 
pca_moment_vars_cov = moment_scaled_cov |>
  select(c(2:1032)) #1:seqn, 1033:age_cat 
pca_moment_cov = prcomp(pca_moment_vars_cov, scale. = TRUE)


# color by systolic BP
fviz_pca_ind(pca_moment_cov, 
             habillage = factor(moment_scaled_cov$sbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by diastolic BP
fviz_pca_ind(pca_moment_cov, 
             habillage = factor(moment_scaled_cov$dbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by total cholesterol
fviz_pca_ind(pca_moment_cov, 
             habillage = factor(moment_scaled_cov$chol_total_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by HDL 
fviz_pca_ind(pca_moment_cov, 
             habillage = factor(moment_scaled_cov$chol_hdl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by LDL 
fviz_pca_ind(pca_moment_cov, 
             habillage = factor(moment_scaled_cov$chol_ldl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by diabetes  
fviz_pca_ind(pca_moment_cov, 
             habillage = factor(moment_scaled_cov$diabetes,
                               levels = 1:3,
                               labels = c("yes", "no", 
                                           "borderline")),  
             label = c("var"))

# color by arthritis  
fviz_pca_ind(pca_moment_cov, 
             habillage = ifelse(moment_scaled_cov$arthritis==1, 
                                "Arthritis", "No"), 
             label = c("var"))

# color by overweight  
fviz_pca_ind(pca_moment_cov, 
             habillage = ifelse(moment_scaled_cov$overweight==1, 
                                "Overweight", "No"), 
             label = c("var"))

# color by malignancy  
fviz_pca_ind(pca_moment_cov, 
             habillage = ifelse(moment_scaled_cov$malig==1, 
                                "Malignancy", "No"), 
             label = c("var"))
```



## 5. PCA of RAW DATA MOMENT embedding 
### 5.1 PCA of RAW DATA MOMENT : color coded by response or demographics 
```{r PCA_Raw MOMENT}
# change age to categories 
data_raw_moment1024 = data_raw_moment1024 |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# define column groups for RAW MOMENT 
raw_moment_embed_cols = c(9:1032)
raw_moment_demo_cols = c("gender", "race", "education", "married", "age") 
raw_moment_response_col = "bmi" # Response variable
raw_moment_id_col = "seqn"  
  

# PCA for RAW MOMENT 
raw_moment_pca_results = pca_analysis(
  data = data_raw_moment1024,
  embed_cols = names(data_raw_moment1024)[raw_moment_embed_cols],
  demo_cols = raw_moment_demo_cols,
  response_col = raw_moment_response_col,
  id_col = raw_moment_id_col
)

raw_moment_pca_results$pca_results$bmi$plot
raw_moment_pca_results$pca_results$gender$plot
raw_moment_pca_results$pca_results$race$plot
raw_moment_pca_results$pca_results$education$plot
raw_moment_pca_results$pca_results$married$plot
raw_moment_pca_results$pca_results$age$plot
```

### 5.2 PCA of RAW DATA MOMENT : color coded by additional covarates 
```{r PCA_Raw MOMENT color coded }
raw_moment_sub_embed = data_raw_moment1024 |>
  select(c(1, 9:1032)) |>
  mutate(across(-seqn, ~ as.numeric(scale(.x))))

raw_moment_sub_cov = data_raw_moment1024[c(1:8, 1033)] |>
  left_join(covariates, by = "seqn")  

raw_moment_scaled_cov = raw_moment_sub_embed |>
  left_join(raw_moment_sub_cov, by = "seqn") |>
  drop_na()


# PCA with embeddings/demographics/bmi of 2914 data points 
pca_raw_moment_vars_cov = raw_moment_scaled_cov |>
  select(c(2:1032)) #1:seqn, 1033:age_cat 
pca_raw_moment_cov = prcomp(pca_raw_moment_vars_cov, scale. = TRUE)


# color by systolic BP
fviz_pca_ind(pca_raw_moment_cov, 
             habillage = factor(raw_moment_scaled_cov$sbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by diastolic BP
fviz_pca_ind(pca_raw_moment_cov, 
             habillage = factor(raw_moment_scaled_cov$dbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by total cholesterol
fviz_pca_ind(pca_raw_moment_cov, 
             habillage = factor(raw_moment_scaled_cov$chol_total_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by HDL 
fviz_pca_ind(pca_raw_moment_cov, 
             habillage = factor(raw_moment_scaled_cov$chol_hdl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by LDL 
fviz_pca_ind(pca_raw_moment_cov, 
             habillage = factor(raw_moment_scaled_cov$chol_ldl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by diabetes  
fviz_pca_ind(pca_raw_moment_cov, 
             habillage = factor(raw_moment_scaled_cov$diabetes,
                               levels = 1:3,
                               labels = c("yes", "no", 
                                           "borderline")),  
             label = c("var"))

# color by arthritis  
fviz_pca_ind(pca_raw_moment_cov, 
             habillage = ifelse(raw_moment_scaled_cov$arthritis==1, 
                                "Arthritis", "No"), 
             label = c("var"))

# color by overweight  
fviz_pca_ind(pca_raw_moment_cov, 
             habillage = ifelse(raw_moment_scaled_cov$overweight==1, 
                                "Overweight", "No"), 
             label = c("var"))

# color by malignancy  
fviz_pca_ind(pca_raw_moment_cov, 
             habillage = ifelse(raw_moment_scaled_cov$malig==1, 
                                "Malignancy", "No"), 
             label = c("var"))
```



## 7. PCA of the 5 min interval data by bmi, gender, race, etc. 
```{r PCA_original 5min interval data}
# import original 5 minute interval data and group by age

# change age to categories 
original_5min = read.csv("./002_data/data_wide.csv") |>
  janitor::clean_names() |>
  select(-x) |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# define column groups for 5min interval data
original_5min_pa_cols = c(9:2024)
original_5min_demo_cols = c("gender", "race", "education", "married", "age") 
original_5min_response_col = "bmi" # Response variable
original_5min_id_col = "seqn"  
  

# PCA for 5min interval data 
original_5min_pca_results = pca_analysis(
  data = original_5min,
  embed_cols = names(original_5min)[original_5min_pa_cols],
  demo_cols = original_5min_demo_cols,
  response_col = original_5min_response_col,
  id_col = original_5min_id_col
)

original_5min_pca_results$pca_results$bmi$plot
original_5min_pca_results$pca_results$gender$plot
original_5min_pca_results$pca_results$race$plot
original_5min_pca_results$pca_results$education$plot
original_5min_pca_results$pca_results$married$plot
original_5min_pca_results$pca_results$age$plot
```



# Visualization - PCA with all covaraites and scaled emebddings  

## 1. PCA of EntroGPT  
```{r PCA_EntroGPT with }
################# Function to do PCA using all covariates 
pca_analysis_allcov = function(data, id_col = "seqn", 
                               exclude_cols = list(), 
                               color_mappings = list()) {
  
  default_mappings = list(
    # all covariates
    bmi = list(
      type = "binary",
      condition = function(x) x == 1,
      true_label = "BMI>25",
      false_label = "Otherwise"),
    gender = list(
      type = "binary",
      condition = function(x) x == 1,
      true_label = "male",
      false_label = "female"),
    race = list(
      type = "categorical",
      levels = 1:5,
      labels = c("Mexican American", "Other Hispanic", 
                 "Non-Hispanic White", 
                "Non-Hispanic Black", "Other")),
    age_cat = list(
      type = "categorical",
      levels = 0:2,
      labels = c("young", "middle", "older")),
    sbp_cat = list(
      type = "categorical",
      levels = 0:3,
      labels = c("low risk", "stage1 hypertension", 
                 "stage2 hypertension", "crisis")),
    dbp_cat = list(
      type = "categorical",
      levels = 0:3,
      labels = c("low risk", "stage1 hypertension", 
                 "stage2 hypertension", "crisis")),
    chol_total_cat = list(
      type = "categorical",
      levels = 0:2,
      labels = c("low risk", "borderline", "high risk")),
    chol_hdl_cat = list(
      type = "categorical",
      levels = 0:2,
      labels = c("low risk", "borderline", "high risk")),
    chol_ldl_cat = list(
      type = "categorical",
      levels = 0:2,
      labels = c("low risk", "borderline", "high risk")),
    diabetes = list(
      type = "categorical",
      levels = 1:3,
      labels = c("yes", "no", "borderline")),
    arthritis = list(
      type = "binary",
      condition = function(x) x == 1,
      true_label = "Arthritis",
      false_label = "No"),
    overweight = list(
      type = "binary",
      condition = function(x) x == 1,
      true_label = "Overweight",
      false_label = "No"),
    malig = list(
      type = "binary",
      condition = function(x) x == 1,
      true_label = "Malignancy",
      false_label = "No")
  )
  
  mappings = default_mappings
  if (length(color_mappings) > 0) {
  for (name in names(color_mappings)) {
    mappings[[name]] = color_mappings[[name]]
  }
}
  
  all_cols = seq_along(names(data))
  
  pca_results = list()
  
  # Process each variable for PCA   
  for (var_name in names(mappings)) {
    
    # exclude id_col = "seqn"
    cols_to_exclude = c(which(names(data) == id_col))
    
    # exclude variable to be color coded 
    if (length(exclude_cols) > 0 && var_name %in% names(exclude_cols)) {
      var_exclusions = exclude_cols[[var_name]]

      if (is.character(var_exclusions)) {
        var_exclusions = which(names(data) %in% var_exclusions)
        }
      cols_to_exclude = c(cols_to_exclude, var_exclusions)
      }
    
    # columns for PCA 
    pca_cols = all_cols[!all_cols %in% cols_to_exclude]
    
    # PCA
    pca_data = data[, pca_cols, drop = FALSE]
    pca_model = prcomp(pca_data, scale. = TRUE)
    
    # Create the plot based on the variable type
    var_mapping = mappings[[var_name]]
    
    if (var_mapping$type == "binary") {
      # Binary 
      plot = fviz_pca_ind(
        pca_model,
        habillage = ifelse(
          var_mapping$condition(data[[var_name]]),
          var_mapping$true_label,
          var_mapping$false_label
        ),
        label = "var",
        title = paste("PCA colored by", var_name)
      )
    } else if (var_mapping$type == "categorical") {
      # Categorical 
      plot = fviz_pca_ind(
        pca_model,
        habillage = factor(
          data[[var_name]],
          levels = var_mapping$levels,
          labels = var_mapping$labels
        ),
        label = "var",
        title = paste("PCA colored by", var_name)
      )
    } else {
      warning(paste("Unknown variable type for", var_name, ". Skipping."))
      next
    }
    
    # results
    pca_results[[var_name]] = list(
      model = pca_model,
      plot = plot,
      excluded_cols = cols_to_exclude
    )
  }
  
  return(pca_results)
}


entrogpt_excl_cols = list(
  bmi = c("seqn", "bmi", "age_cat", 1562:1566), 
  gender = c("seqn", "gender", "age_cat", 1562:1566),
  race = c("seqn", "race", "age_cat", 1562:1566),
  age_cat = c("seqn", "age", "age_cat", 1562:1566),
  sbp_cat = c("seqn", "age_cat", "med_sbp", 1562:1566),
  dbp_cat = c("seqn", "age_cat", "med_dbp", 1562:1566),
  chol_total_cat = c("seqn", "age_cat", "chol_total", 1562:1566),
  chol_hdl_cat = c("seqn", "age_cat", "chol_hdl", 1562:1566),
  chol_ldl_cat = c("seqn", "age_cat", "chol_ldl", 1562:1566),
  diabetes = c("seqn", "age_cat", "diabetes", 1562:1566),
  arthritis = c("seqn", "age_cat", "arthritis", 1562:1566),
  overweight = c("seqn", "age_cat", "overweight", 1562:1566),
  malig = c("seqn", "age_cat", "malig", 1562:1566)
)

# Run the function
pca_entrogpt_allcov_res = pca_analysis_allcov(data = entrogpt_scaled_cov,
                                         id_col = "seqn",
                                         exclude_cols = entrogpt_excl_cols)

# PCA plots
pca_entrogpt_allcov_res$bmi$plot
pca_entrogpt_allcov_res$gender$plot
pca_entrogpt_allcov_res$race$plot
pca_entrogpt_allcov_res$age_cat$plot
pca_entrogpt_allcov_res$sbp_cat$plot
pca_entrogpt_allcov_res$dbp_cat$plot
pca_entrogpt_allcov_res$chol_total_cat$plot
pca_entrogpt_allcov_res$chol_hdl_cat$plot
pca_entrogpt_allcov_res$chol_ldl_cat$plot
pca_entrogpt_allcov_res$diabetes$plot
pca_entrogpt_allcov_res$arthritis$plot
pca_entrogpt_allcov_res$overweight$plot
pca_entrogpt_allcov_res$malig$plot
```


## 2. PCA of EntroBert  
```{r PCA_EntroBERT_with}
entrobert_excl_cols = list(
  bmi = c("seqn", "bmi", "age_cat", 794:798), 
  gender = c("seqn", "gender", "age_cat", 794:798),
  race = c("seqn", "race", "age_cat", 794:798),
  age_cat = c("seqn", "age", "age_cat", 794:798),
  sbp_cat = c("seqn", "age_cat", "med_sbp", 794:798),
  dbp_cat = c("seqn", "age_cat", "med_dbp", 794:798),
  chol_total_cat = c("seqn", "age_cat", "chol_total", 794:798),
  chol_hdl_cat = c("seqn", "age_cat", "chol_hdl", 794:798),
  chol_ldl_cat = c("seqn", "age_cat", "chol_ldl", 794:798),
  diabetes = c("seqn", "age_cat", "diabetes", 794:798),
  arthritis = c("seqn", "age_cat", "arthritis", 794:798),
  overweight = c("seqn", "age_cat", "overweight", 794:798),
  malig = c("seqn", "age_cat", "malig", 794:798)
)

# Run the function
pca_entrobert_allcov_res = pca_analysis_allcov(data = entrobert_scaled_cov,
                                         id_col = "seqn",
                                         exclude_cols = entrobert_excl_cols)

# PCA plots
pca_entrobert_allcov_res$bmi$plot
pca_entrobert_allcov_res$gender$plot
pca_entrobert_allcov_res$race$plot
pca_entrobert_allcov_res$age_cat$plot
pca_entrobert_allcov_res$sbp_cat$plot
pca_entrobert_allcov_res$dbp_cat$plot
pca_entrobert_allcov_res$chol_total_cat$plot
pca_entrobert_allcov_res$chol_hdl_cat$plot
pca_entrobert_allcov_res$chol_ldl_cat$plot
pca_entrobert_allcov_res$diabetes$plot
pca_entrobert_allcov_res$arthritis$plot
pca_entrobert_allcov_res$overweight$plot
pca_entrobert_allcov_res$malig$plot
```


## 3. PCA of EntroCohere  
```{r PCA_EntroCohere_with}
entrocohere_excl_cols = list(
  bmi = c("seqn", "bmi", "age_cat", 1050:1054), 
  gender = c("seqn", "gender", "age_cat", 1050:1054),
  race = c("seqn", "race", "age_cat", 1050:1054),
  age_cat = c("seqn", "age", "age_cat", 1050:1054),
  sbp_cat = c("seqn", "age_cat", "med_sbp", 1050:1054),
  dbp_cat = c("seqn", "age_cat", "med_dbp", 1050:1054),
  chol_total_cat = c("seqn", "age_cat", "chol_total_cat", 1050:1054),
  chol_hdl_cat = c("seqn", "age_cat", "chol_hdl_cat", 1050:1054),
  chol_ldl_cat = c("seqn", "age_cat", "chol_ldl_cat", 1050:1054),
  diabetes = c("seqn", "age_cat", "diabetes", 1050:1054),
  arthritis = c("seqn", "age_cat", "arthritis", 1050:1054),
  overweight = c("seqn", "age_cat", "overweight", 1050:1054),
  malig = c("seqn", "age_cat", "malig", 1050:1054)
)

# Run the function
pca_entrocohere_allcov_res = pca_analysis_allcov(data = entrocohere_scaled_cov,
                                         id_col = "seqn",
                                         exclude_cols = entrocohere_excl_cols)

# PCA plots
pca_entrocohere_allcov_res$bmi$plot
pca_entrocohere_allcov_res$gender$plot
pca_entrocohere_allcov_res$race$plot
pca_entrocohere_allcov_res$age_cat$plot
pca_entrocohere_allcov_res$sbp_cat$plot
pca_entrocohere_allcov_res$dbp_cat$plot
pca_entrocohere_allcov_res$chol_total_cat$plot
pca_entrocohere_allcov_res$chol_hdl_cat$plot
pca_entrocohere_allcov_res$chol_ldl_cat$plot
pca_entrocohere_allcov_res$diabetes$plot
pca_entrocohere_allcov_res$arthritis$plot
pca_entrocohere_allcov_res$overweight$plot
pca_entrocohere_allcov_res$malig$plot
```


## 4. PCA of MOMENT 
```{r PCA_MOMENT_with}
moment_excl_cols = list(
  bmi = c("seqn", "bmi", "age_cat", 1043:1047), 
  gender = c("seqn", "gender", "age_cat", 1043:1047),
  race = c("seqn", "race", "age_cat", 1043:1047),
  age_cat = c("seqn", "age", "age_cat", 1043:1047),
  sbp_cat = c("seqn", "age_cat", "med_sbp", 1043:1047),
  dbp_cat = c("seqn", "age_cat", "med_dbp", 1043:1047),
  chol_total_cat = c("seqn", "age_cat", "chol_total_cat", 1043:1047),
  chol_hdl_cat = c("seqn", "age_cat", "chol_hdl_cat", 1043:1047),
  chol_ldl_cat = c("seqn", "age_cat", "chol_ldl_cat", 1043:1047),
  diabetes = c("seqn", "age_cat", "diabetes", 1043:1047),
  arthritis = c("seqn", "age_cat", "arthritis", 1043:1047),
  overweight = c("seqn", "age_cat", "overweight", 1043:1047),
  malig = c("seqn", "age_cat", "malig", 1043:1047)
)

# Run the function
pca_moment_allcov_res = pca_analysis_allcov(data = moment_scaled_cov,
                                         id_col = "seqn",
                                         exclude_cols = moment_excl_cols)

# PCA plots
pca_moment_allcov_res$bmi$plot
pca_moment_allcov_res$gender$plot
pca_moment_allcov_res$race$plot
pca_moment_allcov_res$age_cat$plot
pca_moment_allcov_res$sbp_cat$plot
pca_moment_allcov_res$dbp_cat$plot
pca_moment_allcov_res$chol_total_cat$plot
pca_moment_allcov_res$chol_hdl_cat$plot
pca_moment_allcov_res$chol_ldl_cat$plot
pca_moment_allcov_res$diabetes$plot
pca_moment_allcov_res$arthritis$plot
pca_moment_allcov_res$overweight$plot
pca_moment_allcov_res$malig$plot
```


## 5. PCA of MOMENT RAW embedding  
```{r PCA_RAW MOMENT_with}
raw_moment_excl_cols = list(
  bmi = c("seqn", "bmi", "age_cat", 1043:1047), 
  gender = c("seqn", "gender", "age_cat", 1043:1047),
  race = c("seqn", "race", "age_cat", 1043:1047),
  age_cat = c("seqn", "age", "age_cat", 1043:1047),
  sbp_cat = c("seqn", "age_cat", "med_sbp", 1043:1047),
  dbp_cat = c("seqn", "age_cat", "med_dbp", 1043:1047),
  chol_total_cat = c("seqn", "age_cat", "chol_total_cat", 1043:1047),
  chol_hdl_cat = c("seqn", "age_cat", "chol_hdl_cat", 1043:1047),
  chol_ldl_cat = c("seqn", "age_cat", "chol_ldl_cat", 1043:1047),
  diabetes = c("seqn", "age_cat", "diabetes", 1043:1047),
  arthritis = c("seqn", "age_cat", "arthritis", 1043:1047),
  overweight = c("seqn", "age_cat", "overweight", 1043:1047),
  malig = c("seqn", "age_cat", "malig", 1043:1047)
)


# Run the function
pca_raw_moment_allcov_res = pca_analysis_allcov(data = raw_moment_scaled_cov,
                                         id_col = "seqn",
                                         exclude_cols = raw_moment_excl_cols)

# PCA plots
pca_raw_moment_allcov_res$bmi$plot
pca_raw_moment_allcov_res$gender$plot
pca_raw_moment_allcov_res$race$plot
pca_raw_moment_allcov_res$age_cat$plot
pca_raw_moment_allcov_res$sbp_cat$plot
pca_raw_moment_allcov_res$dbp_cat$plot
pca_raw_moment_allcov_res$chol_total_cat$plot
pca_raw_moment_allcov_res$chol_hdl_cat$plot
pca_raw_moment_allcov_res$chol_ldl_cat$plot
pca_raw_moment_allcov_res$diabetes$plot
pca_raw_moment_allcov_res$arthritis$plot
pca_raw_moment_allcov_res$overweight$plot
pca_raw_moment_allcov_res$malig$plot
```
