==> 05_embedding_vis2.Rmd <==
---
title: "embedding visualization with PCA"
output: html_document
---

# Embedding Visualization 

```{r setup, include=FALSE}
library(tidyverse)
library(foreign)
library(haven)
library(readr)
library(Rtsne)
library(patchwork)
```

# Data import 
Continued from embedding_evaluating.Rmd 

## 1. Import more covariates 
```{r}
# importing more covariates from 2003-2004
bp_d1 = read_xpt("001_raw_data/BPX_C.xpt") |>
  janitor::clean_names() |>
  rowwise() |>
  mutate(
    med_sbp = median(c(bpxsy1, bpxsy2, bpxsy3, bpxsy4), na.rm = TRUE),
    med_dbp = median(c(bpxdi1, bpxdi2, bpxdi3, bpxdi4), na.rm = TRUE)
  ) |>
  select(seqn, med_sbp, med_dbp) |> # blood pressure in mmHg
  drop_na()

### check if it is okay to use average BP without considering the enhancement usage after each reading? 

ldl_d1 = read_xpt("001_raw_data/L13AM_C.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbdldl)

choles_d1 = read_xpt("001_raw_data/L13_C.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbxtc, lbxhdd) |>
  left_join(ldl_d1, by = "seqn") |>
  rename(chol_total = lbxtc, 
         chol_hdl = lbxhdd, 
         chol_ldl = lbdldl) # mg/dL

diabetes_d1 = read_xpt("001_raw_data/DIQ_C.xpt") |>
  janitor::clean_names() |>
  select(seqn, diq010) |>
  filter(diq010 !=7 & diq010 !=9) |> 
  rename(diabetes = diq010) # 1=yes, 2=no, 3=borderline, 7=refused, 9=dontknow

mc_d1 = read_xpt("001_raw_data/MCQ_C.xpt") |>
  janitor::clean_names() |>
  select(seqn, mcq160a, mcq160j, mcq220) |>
  filter(mcq160a != 7 & mcq160a != 9 & mcq160j != 7 & mcq160j != 9 & 
           mcq220 != 7 & mcq220 != 9) |> 
  rename(arthritis = mcq160a, 
         overweight = mcq160j, 
         malig = mcq220) # 1=yes, 2=no, 7=refused, 9=dontknow 

cov_d1 = bp_d1 |> 
  left_join(choles_d1, by = "seqn") |>
  left_join(diabetes_d1, by = "seqn") |>
  left_join(mc_d1, by ="seqn")


# importing more covariates from 2005-2006
bp_d2 = read_xpt("001_raw_data/BPX_D.xpt") |>
  janitor::clean_names() |>
  rowwise() |>
  mutate(
    med_sbp = median(c(bpxsy1, bpxsy2, bpxsy3, bpxsy4), na.rm = TRUE),
    med_dbp = median(c(bpxdi1, bpxdi2, bpxdi3, bpxdi4), na.rm = TRUE)
  ) |>
  select(seqn, med_sbp, med_dbp) |> #blood pressure in mmHg
  drop_na()

ldl_d2 = read_xpt("001_raw_data/TRIGLY_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbdldl)

hdl_d2 = read_xpt("001_raw_data/HDL_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbdhdd)

choles_d2 = read_xpt("001_raw_data/TCHOL_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, lbxtc) |>
  left_join(ldl_d2, by = "seqn") |>
  left_join(hdl_d2, by = "seqn") |>
  rename(chol_total = lbxtc, 
         chol_hdl = lbdhdd, 
         chol_ldl = lbdldl) # mg/dL

diabetes_d2 = read_xpt("001_raw_data/DIQ_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, diq010) |>
  filter(diq010 !=7 & diq010 !=9) |> 
  rename(diabetes = diq010) # 1=yes, 2=no, 3=borderline, 7=refused, 9=dontknow

mc_d2 = read_xpt("001_raw_data/MCQ_D.xpt") |>
  janitor::clean_names() |>
  select(seqn, mcq160a, mcq220, mcq080) |>
  filter(mcq160a != 7 & mcq160a != 9 & mcq220 != 7 & mcq220 != 9, mcq080!=7 & mcq080!=9) |> 
  rename(arthritis = mcq160a, 
         overweight = mcq080, 
         malig = mcq220) # 1=yes, 2=no, 7=refused, 9=dontknow 

cov_d2 = bp_d2 |> 
  left_join(choles_d2, by = "seqn") |>
  left_join(diabetes_d2, by ="seqn") |>
  left_join(mc_d2, by = "seqn")



# clean up covariates 
covariates = rbind(cov_d1, cov_d2) |>
  mutate(
    sbp_cat = case_when(med_sbp < 130 ~ 0, 
                          med_sbp >= 130 & med_sbp < 140 ~ 1, 
                          med_sbp >= 140 & med_sbp < 180~ 2, 
                          med_sbp >= 180 ~ 3), 
    dbp_cat = case_when(med_dbp < 80 ~ 0, 
                          med_dbp >= 80 & med_dbp < 90 ~ 1, 
                          med_dbp >= 90 & med_dbp < 120 ~ 2, 
                          med_dbp <= 120 ~ 3), 
    chol_total_cat = case_when(chol_total < 200 ~ 0, 
                                 chol_total >= 200 & chol_total< 240 ~ 1, 
                                 chol_total >= 240 ~ 2), 
    chol_ldl_cat = case_when(chol_ldl < 130 ~ 0, 
                               chol_ldl >= 130 & chol_ldl< 160 ~ 1, 
                               chol_ldl >= 160 ~ 2), 
    chol_hdl_cat = case_when(chol_hdl >= 60 ~ 0, 
                               chol_hdl >= 50 & chol_hdl< 60 ~ 1, 
                               chol_hdl < 50 ~ 2))
```

Variables: 

1) med_sbp: median systolic blood pressure in mmHg
2) med_dbp: median diastolic blood pressure in mmHg
3) chol_total: choleterol in mg/dL
4) chol_ldl: LDL in mg/dL
5) chol_hdl: HDL in mg/dL 
6) diabetes: ever told to have diabetes by doctor (1=yes, 2=no, 3=borderline)
7) arthritis: ever told to have arthritis by doctor (1=yes, 2=no)
8) overweight: ever told to be overweight by doctor (1=yes, 2=no)
9) malig: ever told to have cancer of malignancy by doctor (1=yes, 2=no) 

10) sbp_cat: 0=low risk, 1=stage1 hypertension, 2=stage2 hypertension, 3=crisis
11) dbp_cat: 0=low risk, 1=stage1 hypertension, 2=stage2 hypertension, 3=crisis
12) chol_total_cat: 0=low risk, 1=borderline, 2=high risk
13) chol_ldl_cat: 0=low risk, 1=borderline, 2=high risk
14) chol_hdl_cat: 0=low risk, 1=borderline, 2=high risk


## 2. Import embeddings of other models 
```{r embedding_import}
# GPT with 1536 embedding dimension 
data_gpt1536 = read.csv("./003_entro_data/data_wide_embedding_gpt1536.csv") |>
  janitor::clean_names() |>
  dplyr::select(-x,-combined)

data_gpt1536$embedding = gsub("\\[", "", data_gpt1536$embedding)
data_gpt1536$embedding = gsub("\\]", "", data_gpt1536$embedding)

data_gpt1536 = data_gpt1536 |>
  separate(embedding, into = paste0("var", 1:1536), sep = ",\\s*", convert = TRUE) |>
  mutate(across(starts_with("var"), as.numeric))


# BERT with 768 embedding dimension
data_bert768 = read.csv("./003_entro_data/data_wide_embedding_bert768.csv") |>
  janitor::clean_names() |>
  dplyr::select(-x,-combined)

data_bert768$embedding <- gsub("\\[", "", data_bert768$embedding)
data_bert768$embedding <- gsub("\\]", "", data_bert768$embedding)
data_bert768$embedding <- gsub("\n", " ", data_bert768$embedding)

data_bert768 = data_bert768 |>
  mutate(embedding = str_trim(embedding),  # Remove leading and trailing spaces
         embedding = str_replace_all(embedding, "\\s+", " ")) |> # Replace multiple spaces with a single space 
  separate(embedding, into = paste0("var", 1:768), sep = "\\s+", convert = TRUE) |>
  mutate(across(starts_with("var"), as.numeric))


# Cohere with 1024 embedding dimension
data_cohere1024 = read.csv("./003_entro_data/data_wide_embedding_cohere1024.csv") |>
  janitor::clean_names() |>
  dplyr::select(-x,-n_tokens)

data_cohere1024$embedding = gsub("\\[", "", data_cohere1024$embedding) 
data_cohere1024$embedding = gsub("\\]", "", data_cohere1024$embedding)
data_cohere1024$embedding = gsub(",", " ", data_cohere1024$embedding)

data_cohere1024 = data_cohere1024 |> 
  mutate(embedding = str_trim(embedding),  # Remove leading and trailing spaces
         embedding = str_replace_all(embedding, "\\s+", " ")) |> # Replace multiple spaces with a single space
  separate(embedding, into = paste0("var", 1:1024), sep = "\\s+", convert = TRUE) |>
  mutate(across(starts_with("var"), as.numeric))
```

## 3. Import the entropy 
```{r}
# Entropy
data_entropy = read.csv("./003_entro_data/data_wide_entropy.csv") |>
  janitor::clean_names() |>
  dplyr::select(-x) 

# GPT1536 + entropy
data_gpt1536_entropy = data_entropy |>
  dplyr::select(seqn, entropy_day1:entropy_day7) |> 
  inner_join(data_gpt1536, by = "seqn") 

# BERT768 + entropy
data_bert768_entropy = data_entropy |>
  dplyr::select(seqn, entropy_day1:entropy_day7) |> 
  inner_join(data_bert768, by = "seqn") 

# Cohere1024 + entropy
data_cohere1024_entropy = data_entropy |>
  dplyr::select(seqn, entropy_day1:entropy_day7) |> 
  inner_join(data_cohere1024, by = "seqn") 
```

## 4. Get MOMENT embeddings  
```{r}
# MOMENT with 1024 embedding dimension  
data1_moment1024 = read.csv("./002_data/embeddings_moment_subset1_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn) 

data2_moment1024 = read.csv("./002_data/embeddings_moment_subset2_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn) 

data_moment1024 = rbind(data1_moment1024, data2_moment1024) 
```

## 5. Get MOMENT RAW DATA embeddings 
```{r}
# MOMENT raw data embedding with 1024 dimension  
data1_raw_moment1024 = read.csv("./002_data/embeddings_moment_raw_subset1_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn)

data2_raw_moment1024 = read.csv("./002_data/embeddings_moment_raw_subset2_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn) 

data_raw_moment1024 = rbind(data1_raw_moment1024, data2_raw_moment1024) 
```

## 6. Get MOMENT RAW RECODED embeddings 
```{r}
# MOMENT raw recoded embedding with 1024 dimension  
data1_recode_moment1024 = read.csv("./002_data/embeddings_moment_recoded_subset1_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn)

data2_recode_moment1024 = read.csv("./002_data/embeddings_moment_recoded_subset2_1024.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  arrange(seqn) 

data_recode_moment1024 = rbind(data1_recode_moment1024, data2_recode_moment1024)
```




# Visualization - t-SNE using embeddings (not scaled) only 

## 1. t-SNE of EntroGPT embeddings 
```{r t_SNE_EntroGPT, eval=FALSE}
set.seed(1)

# t-SNE color by response (bmi)
tsne_entrogpt_vars1 = data_gpt1536_entropy |>
  select(-seqn, gender, age, race, education, married, pir, bmi, -age_cat)   
tsne_entrogpt1 = Rtsne(as.matrix(tsne_entrogpt_vars1), dims = 2, perplexity = 30)

tsne_entrogpt1_df = data.frame(
  tSNE1 = tsne_entrogpt1$Y[,1],
  tSNE2 = tsne_entrogpt1$Y[,2],
  bmi = ifelse(data_gpt1536_entropy$bmi == 1, "BMI>25", "Otherwise"))

ggplot(tsne_entrogpt1_df, aes(x = tSNE1, y = tSNE2, color = bmi)) +
  geom_point() +
  labs(title = "t-SNE of Embeddings from EntroGPT Colored by BMI",
       x = "t-SNE Dimension 1",
       y = "t-SNE Dimension 2") +
  theme_minimal()
```

```{r t_SNE_function }
############ Function 
tsne_analysis = function(data, embed_cols, demo_cols, response_col, id_col = "seqn", 
                        dims = 2, perplexity = 30, max_iter = 1000, seed = 1,
                        color_mapping = list()) {
  
  get_labels = function(col_name, values, color_mapping) {
    if (is.null(color_mapping[[col_name]])) {
      
      if (col_name == "bmi") {
        return(ifelse(values == 1, "BMI>25", "Otherwise"))
      } else if (col_name == "gender") {
        return(ifelse(values == 1, "male", "female"))
      } else if (col_name == "race") {
        return(factor(values, levels = 1:5,
                     labels = c("Mexican American", "Other Hispanic", 
                               "Non-Hispanic White", 
                               "Non-Hispanic Black", "Other")))
      } else if (col_name == "education") {
        return(factor(values, levels = 1:5,
                     labels = c("less than 9th grade", "9-11th grade", 
                               "high school grad/GED or equivalent", 
                               "some college or associates degree", 
                               "college graduate or higher")))
      } else if (col_name == "married") {
        return(factor(values, levels = 1:6,
                     labels = c("married", "widowed", "divorced", 
                                "separated", "never married", 
                                "living with partner")))
      } else if (col_name == "age_cat") {
        return(factor(values, levels = 0:2,
                     labels = c("young", "middle", "older")))
      } else if (col_name == "sbp_cat") {
        return(factor(values, levels = 0:3,
               labels = c("low risk", "stage1 hypertension", 
                         "stage2 hypertension", "crisis")))
      } else if (col_name == "dbp_cat") {
        return(factor(values, levels = 0:3,
               labels = c("low risk", "stage1 hypertension", 
                         "stage2 hypertension", "crisis")))
      } else if (col_name == "chol_total_cat") {
        return(factor(values, levels = 0:2,
               labels = c("low risk", "borderline", "high risk")))
      } else if (col_name == "chol_ldl_cat") {
        return(factor(values, levels = 0:2,
               labels = c("low risk", "borderline", "high risk")))
      } else if (col_name == "chol_hdl_cat") {
        return(factor(values, levels = 0:2,
               labels = c("low risk", "borderline", "high risk")))
      } else if (col_name == "diabetes") {
        return(factor(values, levels = 1:3,
               labels = c("yes", "no", "borderline")))
      } else if (col_name == "arthritis") {
        return(ifelse(values == 1, "Arthritis", "No"))
      } else if (col_name == "overweight") {
        return(ifelse(values == 1, "Overweight", "No"))
      } else if (col_name == "malig") {
        return(ifelse(values == 1, "Malignancy", "No"))
      } else {
        return(as.factor(values))
      }
      
    } else {
      mapping = color_mapping[[col_name]]
      if (is.function(mapping)) {
        return(mapping(values))
      } else {
        return(mapping[values])
      }
    }
  }
  
  
  # Separate into embeddings and covariates/response
  embed_data = data[, embed_cols, drop = FALSE]
  covar_data = data[, c(id_col, demo_cols, response_col), drop = FALSE]
  
  if ("age_cat" %in% names(data) && !("age_cat" %in% demo_cols)) {
    covar_data$age_cat = data$age_cat
  }
  
  set.seed(seed)
  
  # Run t-SNE
  tsne_model = Rtsne(as.matrix(embed_data), 
                     dims = dims, 
                     perplexity = perplexity,
                     max_iter = max_iter,
                     check_duplicates = FALSE)
  
  # Create t-SNE results dataframe
  tsne_coords = data.frame(tsne_model$Y)
  colnames(tsne_coords) = paste0("tSNE", 1:dims)
  tsne_coords[[id_col]] = data[[id_col]]
  
  # Combine t-SNE coordinates with covariates
  tsne_data = tsne_coords |>
    left_join(covar_data, by = id_col)
  
  # Create plots
  tsne_results = list()
  
  # t-SNE plot for response variable
  tsne_data[[paste0(response_col, "_labeled")]] = 
    get_labels(response_col, tsne_data[[response_col]], color_mapping)

  tsne_results[[response_col]] = list(
    data = tsne_data,
    plot = ggplot(tsne_data, aes(x = tSNE1, y = tSNE2, 
                                color = .data[[paste0(response_col, "_labeled")]])) + 
      geom_point(alpha = 0.7, size = 1.5) +
      labs(title = paste("t-SNE colored by", response_col),
           x = "t-SNE Dimension 1",
           y = "t-SNE Dimension 2",
           color = response_col) +
      theme_minimal() +
      theme(legend.position = "bottom")
  )
  
  # t-SNE plots for each demographic variable
  for (demo_col in demo_cols) {
    color_col = demo_col
    if (demo_col == "age" && "age_cat" %in% names(tsne_data)) {
      color_col = "age_cat"
    }
    
    label_col_name = paste0(demo_col, "_labeled")
    tsne_data[[label_col_name]] = get_labels(color_col, tsne_data[[color_col]], color_mapping)

    tsne_results[[demo_col]] = list(
      data = tsne_data,
      plot = ggplot(tsne_data, aes(x = tSNE1, y = tSNE2, 
                                  color = .data[[label_col_name]])) +
        geom_point(alpha = 0.7, size = 1.5) +
        labs(title = paste("t-SNE colored by", ifelse(color_col == "age_cat" && demo_col == "age", 
                                                     "age (categorized)", color_col)),
             x = "t-SNE Dimension 1",
             y = "t-SNE Dimension 2",
             color = ifelse(color_col == "age_cat" && demo_col == "age", 
                           "age (categorized)", color_col)) +
        theme_minimal() +
        theme(legend.position = "bottom")
    )
  }
  
  # Return results
  return(list(
    tsne_data = tsne_data,
    tsne_model = tsne_model,
    tsne_results = tsne_results,
    parameters = list(
      dims = dims,
      perplexity = perplexity,
      max_iter = max_iter,
      seed = seed
    )
  ))
}


# covariates 
data_gpt1536_entropy = data_gpt1536_entropy |>
  mutate(age_cat = case_when(age >= 20 & age <= 39 ~ 0,
                             age >= 40 & age <= 64 ~ 1,
                             age >= 65             ~ 2)) |>
  left_join(covariates, by = "seqn") |>
  drop_na()

# Define column groups 
entrogpt_embed_cols = c(2:8, 16:1551)
entrogpt_demo_cols = c("gender", "race", "education", "married", "age", "sbp_cat",
                       "dbp_cat", "chol_total_cat", "chol_ldl_cat", "chol_hdl_cat",
                       "diabetes", "arthritis", "overweight", "malig") 
entrogpt_response_col = "bmi"
entrogpt_id_col = "seqn"

# Run t-SNE for EntroGPT
entrogpt_tsne_results = tsne_analysis(
  data = data_gpt1536_entropy,
  embed_cols = names(data_gpt1536_entropy)[entrogpt_embed_cols],
  demo_cols = entrogpt_demo_cols,
  response_col = entrogpt_response_col,
  id_col = entrogpt_id_col,
  perplexity = 30,
  seed = 1
)

# plots
entrogpt_tsne_results$tsne_results$bmi$plot
entrogpt_tsne_results$tsne_results$gender$plot
entrogpt_tsne_results$tsne_results$race$plot
  
entrogpt_tsne_results$tsne_results$education$plot
entrogpt_tsne_results$tsne_results$married$plot
entrogpt_tsne_results$tsne_results$age$plot
  
entrogpt_tsne_results$tsne_results$sbp_cat$plot
entrogpt_tsne_results$tsne_results$dbp_cat$plot
entrogpt_tsne_results$tsne_results$chol_total_cat$plot

entrogpt_tsne_results$tsne_results$chol_ldl_cat$plot
entrogpt_tsne_results$tsne_results$chol_hdl_cat$plot
entrogpt_tsne_results$tsne_results$diabetes$plot
  
entrogpt_tsne_results$tsne_results$arthritis$plot
entrogpt_tsne_results$tsne_results$overweight$plot
entrogpt_tsne_results$tsne_results$malig$plot

```


## 2. t_SNE of EntroBERT  
```{r t_SNE_EntroBERT, eval=FALSE}
set.seed(1) 

# t-SNE color by response (bmi)
tsne_entrobert_vars1 = data_bert768_entropy |>
  select(-seqn, -age_cat, -bmi)   
tsne_entrobert1 = Rtsne(as.matrix(tsne_entrobert_vars1), dims = 2, perplexity = 30)

tsne_entrobert1_df = data.frame(
  tSNE1 = tsne_entrobert1$Y[,1],
  tSNE2 = tsne_entrobert1$Y[,2],
  bmi = ifelse(data_bert768_entropy$bmi == 1, "BMI>25", "Otherwise"))

ggplot(tsne_entrobert1_df, aes(x = tSNE1, y = tSNE2, color = bmi)) +
  geom_point() +
  labs(title = "t-SNE of Embeddings from EntroBERT Colored by BMI",
       x = "t-SNE Dimension 1",
       y = "t-SNE Dimension 2") +
  theme_minimal()
```

```{r}
############ Function for EntroBERT
data_bert768_entropy = data_bert768_entropy |>
  mutate(age_cat = case_when(age >= 20 & age <= 39 ~ 0,
                             age >= 40 & age <= 64 ~ 1,
                             age >= 65             ~ 2)) |>
  left_join(covariates, by = "seqn") |>
  drop_na()

# Define column groups 
entrobert_embed_cols = c(2:8, 16:783)
entrobert_demo_cols = c("gender", "race", "education", "married", "age", "sbp_cat",
                       "dbp_cat", "chol_total_cat", "chol_ldl_cat", "chol_hdl_cat",
                       "diabetes", "arthritis", "overweight", "malig") 
entrobert_response_col = "bmi"
entrobert_id_col = "seqn"

# Run t-SNE for EntroBERT
entrobert_tsne_results = tsne_analysis(
  data = data_bert768_entropy,
  embed_cols = names(data_bert768_entropy)[entrobert_embed_cols],
  demo_cols = entrobert_demo_cols,
  response_col = entrobert_response_col,
  id_col = entrobert_id_col,
  perplexity = 30,
  seed = 1
)

# plots

  entrobert_tsne_results$tsne_results$bmi$plot 
  entrobert_tsne_results$tsne_results$gender$plot 
  entrobert_tsne_results$tsne_results$race$plot 
  
  entrobert_tsne_results$tsne_results$education$plot
  entrobert_tsne_results$tsne_results$married$plot
  entrobert_tsne_results$tsne_results$age$plot

  entrobert_tsne_results$tsne_results$sbp_cat$plot 
  entrobert_tsne_results$tsne_results$dbp_cat$plot 
  entrobert_tsne_results$tsne_results$chol_total_cat$plot

  entrobert_tsne_results$tsne_results$chol_ldl_cat$plot 
  entrobert_tsne_results$tsne_results$chol_hdl_cat$plot 
  entrobert_tsne_results$tsne_results$diabetes$plot 

  entrobert_tsne_results$tsne_results$arthritis$plot 
  entrobert_tsne_results$tsne_results$overweight$plot 
  entrobert_tsne_results$tsne_results$malig$plot

```

## 3. t_SNE of EntroCohere  
```{r t_SNE_EntroCohere, eval=FALSE}
set.seed(1) 

# t-SNE color by response (bmi)
tsne_entrocohere_vars1 = data_cohere1024_entropy |>
  select(-seqn, -age_cat, -bmi)   
tsne_entrocohere1 = Rtsne(as.matrix(tsne_entrocohere_vars1), dims = 2, perplexity = 30)

tsne_entrocohere1_df = data.frame(
  tSNE1 = tsne_entrocohere1$Y[,1],
  tSNE2 = tsne_entrocohere1$Y[,2],
  bmi = ifelse(data_cohere1024_entropy$bmi == 1, "BMI>25", "Otherwise"))

ggplot(tsne_entrocohere1_df, aes(x = tSNE1, y = tSNE2, color = bmi)) +
  geom_point() +
  labs(title = "t-SNE of Embeddings from EntroCohere Colored by BMI",
       x = "t-SNE Dimension 1",
       y = "t-SNE Dimension 2") +
  theme_minimal()
```

```{r}
############ Function for EntroCohere
data_cohere1024_entropy = data_cohere1024_entropy |>
  mutate(age_cat = case_when(age >= 20 & age <= 39 ~ 0,
                             age >= 40 & age <= 64 ~ 1,
                             age >= 65             ~ 2)) |>
  left_join(covariates, by = "seqn") |>
  drop_na()

# Define column groups 
entrocohere_embed_cols = c(2:8, 16:1039)
entrocohere_demo_cols = c("gender", "race", "education", "married", "age", "sbp_cat",
                       "dbp_cat", "chol_total_cat", "chol_ldl_cat", "chol_hdl_cat",
                       "diabetes", "arthritis", "overweight", "malig") 
entrocohere_response_col = "bmi"
entrocohere_id_col = "seqn"

# Run t-SNE for EntroCohere
entrocohere_tsne_results = tsne_analysis(
  data = data_cohere1024_entropy,
  embed_cols = names(data_cohere1024_entropy)[entrocohere_embed_cols],
  demo_cols = entrocohere_demo_cols,
  response_col = entrocohere_response_col,
  id_col = entrocohere_id_col,
  perplexity = 30,
  seed = 1
)

# plots
  entrocohere_tsne_results$tsne_results$bmi$plot 
  entrocohere_tsne_results$tsne_results$gender$plot 
  entrocohere_tsne_results$tsne_results$race$plot 
    
  entrocohere_tsne_results$tsne_results$education$plot 
  entrocohere_tsne_results$tsne_results$married$plot 
  entrocohere_tsne_results$tsne_results$age$plot 
    
  entrocohere_tsne_results$tsne_results$sbp_cat$plot 
  entrocohere_tsne_results$tsne_results$dbp_cat$plot 
  entrocohere_tsne_results$tsne_results$chol_total_cat$plot 
  
  entrocohere_tsne_results$tsne_results$chol_ldl_cat$plot 
  entrocohere_tsne_results$tsne_results$chol_hdl_cat$plot 
  entrocohere_tsne_results$tsne_results$diabetes$plot 
  
  entrocohere_tsne_results$tsne_results$arthritis$plot 
  entrocohere_tsne_results$tsne_results$overweight$plot 
  entrocohere_tsne_results$tsne_results$malig$plot

```


## 4. t_SNE of MOMENT 
```{r t_SNE_MOMENT, eval=FALSE}
set.seed(1) 

data_moment1024 = data_moment1024 |>
  mutate(age_cat = case_when(age >= 20 & age <= 39 ~ 0,
                             age >= 40 & age <= 64 ~ 1,
                             age >= 65             ~ 2))

# t-SNE color by response (bmi)
tsne_moment_vars1 = data_moment1024 |>
  select(-seqn, bmi, -age_cat)    
tsne_moment1 = Rtsne(as.matrix(tsne_moment_vars1), dims = 2, perplexity = 30)

tsne_moment1_df = data.frame(
  tSNE1 = tsne_moment1$Y[,1],
  tSNE2 = tsne_moment1$Y[,2],
  bmi = ifelse(data_moment1024$bmi == 1, "BMI>25", "Otherwise"))

ggplot(tsne_moment1_df, aes(x = tSNE1, y = tSNE2, color = bmi)) +
  geom_point() +
  labs(title = "t-SNE of Embeddings from MOMENT Colored by BMI",
       x = "t-SNE Dimension 1",
       y = "t-SNE Dimension 2") +
  theme_minimal()

# t-SNE color by age_cat 
tsne_moment_vars2 = data_moment1024 |>
  select(-seqn, age, -age_cat)   
tsne_moment2 = Rtsne(as.matrix(tsne_moment_vars2), dims = 2, perplexity = 30)

tsne_moment2_df = data.frame(
  tSNE1 = tsne_moment2$Y[,1],
  tSNE2 = tsne_moment2$Y[,2],
  age_cat = factor(data_moment1024$age_cat, levels = 1:3, labels = c("young", "middle", "older")))

ggplot(tsne_moment2_df, aes(x = tSNE1, y = tSNE2, color = age_cat)) +
  geom_point() +
  labs(title = "t-SNE of Embeddings from MOMENT Colored by age_cat",
       x = "t-SNE Dimension 1",
       y = "t-SNE Dimension 2") +
  theme_minimal()
```

```{r}
############ Function for MOMENT
data_moment1024 = data_moment1024 |>
  mutate(age_cat = case_when(age >= 20 & age <= 39 ~ 0,
                             age >= 40 & age <= 64 ~ 1,
                             age >= 65             ~ 2)) |>
  left_join(covariates, by = "seqn") |>
  drop_na()

# Define column groups 
moment_embed_cols = c(2:8, 16:1032)
moment_demo_cols = c("gender", "race", "education", "married", "age", "sbp_cat",
                       "dbp_cat", "chol_total_cat", "chol_ldl_cat", "chol_hdl_cat",
                       "diabetes", "arthritis", "overweight", "malig") 
moment_response_col = "bmi"
moment_id_col = "seqn"

# Run t-SNE for MOMENT
moment_tsne_results = tsne_analysis(
  data = data_moment1024,
  embed_cols = names(data_moment1024)[moment_embed_cols],
  demo_cols = moment_demo_cols,
  response_col = moment_response_col,
  id_col = moment_id_col,
  perplexity = 30,
  seed = 1
)

# plots

  moment_tsne_results$tsne_results$bmi$plot 
  moment_tsne_results$tsne_results$gender$plot 
  moment_tsne_results$tsne_results$race$plot 
    
  moment_tsne_results$tsne_results$education$plot 
  moment_tsne_results$tsne_results$married$plot 
  moment_tsne_results$tsne_results$age$plot 
    
  moment_tsne_results$tsne_results$sbp_cat$plot 
  moment_tsne_results$tsne_results$dbp_cat$plot 
  moment_tsne_results$tsne_results$chol_total_cat$plot 
    
  moment_tsne_results$tsne_results$chol_ldl_cat$plot 
  moment_tsne_results$tsne_results$chol_hdl_cat$plot
  moment_tsne_results$tsne_results$diabetes$plot 
    
  moment_tsne_results$tsne_results$arthritis$plot 
  moment_tsne_results$tsne_results$overweight$plot 
  moment_tsne_results$tsne_results$malig$plot

```


## 5. t_SNE of RAW MOMENT embeddings 
```{r}
############ Function for RAW MOMENT
data_raw_moment1024 = data_raw_moment1024 |>
  mutate(age_cat = case_when(age >= 20 & age <= 39 ~ 0,
                             age >= 40 & age <= 64 ~ 1,
                             age >= 65             ~ 2)) |>
  left_join(covariates, by = "seqn") |>
  drop_na()

# Define column groups 
raw_moment_embed_cols = c(2:8, 16:1039)
raw_moment_demo_cols = c("gender", "race", "education", "married", "age", "sbp_cat",
                       "dbp_cat", "chol_total_cat", "chol_ldl_cat", "chol_hdl_cat",
                       "diabetes", "arthritis", "overweight", "malig") 
raw_moment_response_col = "bmi"
raw_moment_id_col = "seqn"

# Run t-SNE for EntroCohere
raw_moment_tsne_results = tsne_analysis(
  data = data_raw_moment1024,
  embed_cols = names(data_raw_moment1024)[raw_moment_embed_cols],
  demo_cols = raw_moment_demo_cols,
  response_col = raw_moment_response_col,
  id_col = raw_moment_id_col,
  perplexity = 30,
  seed = 1
)

# plots

  raw_moment_tsne_results$tsne_results$bmi$plot 
  raw_moment_tsne_results$tsne_results$gender$plot 
  raw_moment_tsne_results$tsne_results$race$plot 
    
  raw_moment_tsne_results$tsne_results$education$plot 
  raw_moment_tsne_results$tsne_results$married$plot 
  raw_moment_tsne_results$tsne_results$age$plot 

  raw_moment_tsne_results$tsne_results$sbp_cat$plot 
  raw_moment_tsne_results$tsne_results$dbp_cat$plot
  raw_moment_tsne_results$tsne_results$chol_total_cat$plot 
    
  raw_moment_tsne_results$tsne_results$chol_ldl_cat$plot 
  raw_moment_tsne_results$tsne_results$chol_hdl_cat$plot 
  raw_moment_tsne_results$tsne_results$diabetes$plot 
    
  raw_moment_tsne_results$tsne_results$arthritis$plot 
  raw_moment_tsne_results$tsne_results$overweight$plot 
  raw_moment_tsne_results$tsne_results$malig$plot

  
```







# Visualization - Clustering
## 1. K-means clustering of the original data 
```{r eval=FALSE}
fviz_nbclust(original_5min, 
             FUNcluster = kmeans, 
             method = "silhouette")

# optimal k = 2 

km_original_5min = kmeans(original_5min, center = 2, nstart = 20)

km_original_vis = 
  fviz_cluster(list(data = original_5min, cluster = km_original_5min$cluster), 
               geom = c("point", "text"), 
               labelsize = 5, 
               palette = "Dark2") + 
  labs(title = "K-means of the original 5min data")
km_original_vis
```

## 2. K-means clustering of the moment embedding 
```{r eval=FALSE}
fviz_nbclust(data_moment1024, 
             FUNcluster = kmeans, 
             method = "silhouette")

# optimal k = 2 

km_moment = kmeans(data_moment1024, center = 2, nstart = 20)

km_moment_vis = 
  fviz_cluster(list(data = data_moment1024, cluster = km_moment$cluster), 
               geom = c("point", "text"), 
               labelsize = 5, 
               palette = "Dark2") + 
  labs(title = "K-means of the moment embedding")
km_moment_vis
```
