==> 12_entropy.ipynb <==
!pip install nolds

import nolds
import numpy as np
import pandas as pd
from scipy.stats import entropy

df = pd.read_csv("./data/data_wide.csv", index_col=0)
df.head(3)

# Initialize placeholders for the entropy values of each day
entropy_day1 = []
entropy_day2 = []
entropy_day3 = []
entropy_day4 = []
entropy_day5 = []
entropy_day6 = []
entropy_day7 = []

# Iterate over each row in the DataFrame
for index, row in df.iterrows():
    # Extract step data for each individual
    steps_per_day = [row[f'time{i}'] for i in range(1, 2017)]
    daily_entropy = []
    for day in range(7):
        # Get steps data for each day (288 min5 per day)
        steps = steps_per_day[day * 288:(day + 1) * 288]
        value, counts = np.unique(steps, return_counts=True)
        probability_distribution = counts / len(steps)
        
        # Calculate Shannon entropy for the day
        shannon_entropy = entropy(probability_distribution, base=2)
        daily_entropy.append(shannon_entropy)
    
    # Append the entropy values for each day
    entropy_day1.append(daily_entropy[0])
    entropy_day2.append(daily_entropy[1])
    entropy_day3.append(daily_entropy[2])
    entropy_day4.append(daily_entropy[3])
    entropy_day5.append(daily_entropy[4])
    entropy_day6.append(daily_entropy[5])
    entropy_day7.append(daily_entropy[6])

# Add the entropy values for each day to the DataFrame
df['Entropy_Day1'] = entropy_day1
df['Entropy_Day2'] = entropy_day2
df['Entropy_Day3'] = entropy_day3
df['Entropy_Day4'] = entropy_day4
df['Entropy_Day5'] = entropy_day5
df['Entropy_Day6'] = entropy_day6
df['Entropy_Day7'] = entropy_day7

df = df[[col for col in df.columns if not col.startswith('time')]]

df.head(3)

df.to_csv("./data/data_wide_entropy.csv")
df.head(3)
