==> 99_drafts.Rmd <==
---
title: "embedding_evaluating_draft"
author: "Soomin You"
output: html_document
---


## Calculate AUC using EntroLLM method. Try moment first 
```{r}
set.seed(1)

split_moment = sample(nrow(data_moment50), floor(nrow(data_moment50)*0.8), replace = FALSE)
train_m = data_moment50[split_moment,]
test_m = data_moment50[-split_moment,]

train_m_x = as.matrix(train_m[, !(names(train_m) %in% c("bmi", "seqn"))])
train_m_y = train_m$bmi
test_m_x = as.matrix(test_m[, !(names(test_m) %in% c("bmi", "seqn"))])
test_m_y = test_m$bmi 
    
cv_ridge_m = cv.glmnet(train_m_x, train_m_y, alpha = 0, family = 'binomial',
                          standardize = TRUE, grouped = FALSE, nfolds = 10)
coef_m = predict(cv_ridge_m, s = "lambda.min", type = "coefficients")
    
pred_pro_ridge_m = predict(cv_ridge_m, s = cv_ridge_m$lambda.min, newx = test_m_x)
    
pred_ridge_m = exp(pred_pro_ridge_m)/(1+exp(pred_pro_ridge_m))
pred_ridge_m = as.numeric(pred_ridge_m)
    
auc_ridge_m = auc(test_m_y, pred_ridge_m)
auc_m = data.frame(auc_ridge = auc_ridge_m)

roc_obj_m = pROC::roc(test_m_y, pred_ridge_m)
plot(roc_obj_m, col = "blue", main = paste("ROC Curve (AUC =", round(auc_ridge_m, 3), ")"))
```


## Compare AUC with other models 
```{r}
# simulation
model_auc = function(data, sim){ 

  auc = data.frame()
  for (i in 1:sim){
    
    set.seed(i)
    select_train = sample(nrow(data), floor(nrow(data)*0.8), replace = FALSE)
    train = data[select_train,]
    test = data[-select_train,]
    
    train_x = as.matrix(train[, !(names(train) %in% c("bmi", "seqn"))])
    train_y = train$bmi
    test_x = as.matrix(test[, !(names(test) %in% c("bmi", "seqn"))])
    test_y = test$bmi  
    
    cv.ridge = cv.glmnet(train_x,train_y,alpha = 0,family = 'binomial',
                          standardize=TRUE,grouped=FALSE,nfolds = 10)
    coef_ridge=predict(cv.ridge, s = "lambda.min", type = "coefficients")
    
    y_predicted_pro_ridge = predict(cv.ridge,s=cv.ridge$lambda.min,newx=test_x)
    
    y_predicted_ridge = exp(y_predicted_pro_ridge)/(1+exp(y_predicted_pro_ridge))
    y_predicted_ridge = as.numeric(y_predicted_ridge)
    
    auc_ridge = auc(test_y, y_predicted_ridge)
    
    inter.auc = data.frame(auc_ridge = auc_ridge)
    auc = rbind(auc, inter.auc) 
  }  
  return(auc=auc)
}

auc_demo_only = model_auc(data_demo_only, sim = 10)
auc_base = model_auc(data_wide, sim = 10)
#auc_EntroGPT1536 = model_auc(data_gpt1536_entropy, sim = 10)
auc_EntroGPT50 = model_auc(data_gpt50_entropy, sim = 10)
auc_EntroBERT768 = model_auc(data_bert768_entropy, sim = 10)
auc_EntroBERT50 = model_auc(data_bert50_entropy, sim = 10)
auc_EntroCohere1024 = model_auc(data_cohere1024_entropy, sim = 10)
auc_EntroCohere50 = model_auc(data_cohere50_entropy, sim = 10)
auc_MOMENT1024 = model_auc(data_moment1024, sim = 10)
auc_MOMENT50 = model_auc(data_moment50, sim = 10)

auc_table = data.frame(
  Model = c("Demographic Only", "Baseline", "EntroGPT50", "EntroBERT768", 
                 "EntroBERT50", "EntroCohere1024", "EntroCohere50", "MOMENT1024",
            "MOMENT50"),
  AUC = c(mean(auc_demo_only$auc_ridge), mean(auc_base$auc_ridge),
          mean(auc_EntroGPT50$auc_ridge), 
          mean(auc_EntroBERT768$auc_ridge), mean(auc_EntroBERT50$auc_ridge),
          mean(auc_EntroCohere1024$auc_ridge), mean(auc_EntroCohere50$auc_ridge), 
          mean(auc_MOMENT1024$auc_ridge), mean(auc_MOMENT50$auc_ridge)))
auc_table
```


## Plot for comparison 
```{r} 
# plot AUC 

AUC_ridge = cbind(auc_demo_only, auc_base, auc_EntroGPT50, 
                  auc_EntroBERT768, auc_EntroBERT50, 
                  auc_EntroCohere1024, auc_EntroCohere50, 
                  auc_MOMENT1024, auc_MOMENT50)

colnames(AUC_ridge) = c("Demographic Only", "Baseline", "EntroGPT50", 
                        "EntroBERT768", "EntroBERT50", 
                        "EntroCohere1024", "EntroCohere50", 
                        "MOMENT1024", "MOMENT50")

AUC_ridge = AUC_ridge |>
  mutate(across(everything(), as.numeric)) |> 
  pivot_longer(cols = everything(),  
               names_to = "Model",  
               values_to = "AUC") 

AUC_ridge$Model = factor(AUC_ridge$Model, 
                         levels = c("Demographic Only", "Baseline", "EntroGPT50", 
                        "EntroBERT768", "EntroBERT50", 
                        "EntroCohere1024", "EntroCohere50", 
                        "MOMENT1024", "MOMENT50"))

ggplot(AUC_ridge, aes(x = Model, y = AUC, fill = Model)) +
  geom_boxplot() +
  theme(
    legend.text = element_text(hjust = 0),
    axis.text.x = element_text(angle = 45, hjust = 1),
    legend.spacing.x = unit(2, "cm"),
    panel.grid.major = element_line(colour = "grey80", size = 0.5),  
    panel.grid.minor = element_line(colour = "grey90", size = 0.25)  
  )  


######## ROC plot.
model_f1 = function(data, seed, interval) {
  
  set.seed(seed)
  select_train = sample(nrow(data), floor(nrow(data)*0.8), replace = FALSE)
  train = data[select_train,]
  test = data[-select_train,]
  
  train_x = as.matrix(train[, !(names(train) %in% c("bmi", "seqn"))])
  train_y = train$bmi
  test_x = as.matrix(test[, !(names(test) %in% c("bmi", "seqn"))])
  test_y = test$bmi 
  
  cv.fit = cv.glmnet(train_x, train_y, alpha = 0, family = 'binomial',
                      standardize=TRUE,grouped=FALSE,nfolds = 10)
  # use ridge: alpha = 0 
  # coef_ridge=predict(cv.ridge, s = "lambda.min", type = "coefficients")
  
  y_predicted_pro = predict(cv.fit, s = cv.fit$lambda.min, newx = test_x)
  
  y_predicted = exp(y_predicted_pro)/(1+exp(y_predicted_pro))
  y_predicted = as.numeric(y_predicted)
  
  roc = pROC::roc(test_y, y_predicted)
  
  # Initialize variables to store the best cutoff and the best F1 score
  best_cutoff = 0
  best_f1 = 0
  
  # Try different cutoff values to find the one that maximizes the F1 score
  for (cutoff in seq(0.5, 1, by = interval)) {
    y_hat = ifelse(y_predicted >= cutoff, 1, 0)
    f1 = F1_Score(y_hat, test_y)
    if (!is.na(f1) && f1 > best_f1) {
      best_f1 <- f1
      best_cutoff <- cutoff
    }
  }
  
  # calculate true positive, false positive, precision and recall based on maximum F1 score
  maxF1_y_hat = ifelse(y_predicted >= best_cutoff, 1, 0)
  maxF1_confusion = confusionMatrix(factor(maxF1_y_hat, levels=c(0, 1)), 
                                    factor(test_y, levels=c(0, 1)))$table
  maxF1_precision = maxF1_confusion[2, 2] / sum(maxF1_confusion[2, ])
  maxF1_recall = maxF1_confusion[2, 2] / sum(maxF1_confusion[, 2])
  maxF1_TP = maxF1_confusion[2, 2]
  maxF1_FP = maxF1_confusion[2, 1]
  
  return(list(roc = roc, best_cutoff = best_cutoff, best_f1 = best_f1, 
              maxF1_precision = maxF1_precision, maxF1_recall = maxF1_recall, 
              maxF1_TP = maxF1_TP, maxF1_FP = maxF1_FP))
}


F1_Score = function(y_hat,y_true) {
  confusion = confusionMatrix(factor(y_hat, levels = c(0, 1)), 
                              factor(y_true, levels=c(0, 1)))$table
  precision = confusion[2, 2] / sum(confusion[2, ])
  recall = confusion[2, 2] / sum(confusion[, 2])
  f1 = 2 * (precision * recall) / (precision + recall)
  return(f1)
}

f1_demo_only = model_f1(data_demo_only, seed=1, interval=0.01)
f1_base = model_f1(data_wide, seed=1, interval=0.01)
#f1_EntroGPT1536 <- model_f1(data_gpt1536_entropy, seed=1, interval=0.01)
f1_EntroGPT50 = model_f1(data_gpt50, seed=1, interval=0.01)
f1_EntroBERT768 = model_f1(data_bert768, seed=1, interval=0.01)
f1_EntroBERT50 <- model_f1(data_bert50, seed=1, interval=0.01)
f1_EntroCohere1024 <- model_f1(data_cohere1024, seed=1, interval=0.01)
f1_EntroCohere50 <- model_f1(data_cohere50, seed=1, interval=0.01)
f1_MOMENT1024 <- model_f1(data_moment1024, seed=1, interval=0.01)
f1_MOMENT50 <- model_f1(data_moment50, seed=1, interval=0.01)

f1_file_names <- list(
  f1_demo_only = "f1_demo_only",
  f1_base = "f1_base",
  f1_EntroGPT50 = "f1_EntroGPT50",
  f1_EntroBERT768 = "f1_EntroBERT768",
  f1_EntroBERT50 = "f1_EntroBERT50",
  f1_EntroCohere1024 = "f1_EntroCohere1024",
  f1_EntroCohere50 = "f1_EntroCohere50", 
  f1_MOMENT1024 = "f1_MOMENT1024", 
  f1_MOMENT50 = "f1_MOMENT50"
)

roc <- function(Model, filename) {
  data.frame(
    specificity = Model[["roc"]][["specificities"]],
    sensitivity = Model[["roc"]][["sensitivities"]],
    Model = filename)
}

roc_list <- list()

for (var_name in f1_file_names) {
  model_data <- get(var_name)
  filename_prefix <- sub("^f1_", "", var_name)
  filename <- filename_prefix
  roc_list[[filename]] <- roc(model_data, filename_prefix)
}

new_model_names <- c("Demographic Only", "Baseline", "EntroGPT50", 
                        "EntroBERT768", "EntroBERT50", 
                        "EntroCohere1024", "EntroCohere50", 
                        "MOMENT1024", "MOMENT50")

names(roc_list) <- new_model_names 

for (i in seq_along(roc_list)) {
  roc_list[[i]]$Model <- new_model_names[i]
}

roc_data <- do.call(rbind, roc_list)

roc_data$Model <- factor(roc_data$Model, 
                         levels = c("Demographic Only", "Baseline", "EntroGPT50", 
                        "EntroBERT768", "EntroBERT50", 
                        "EntroCohere1024", "EntroCohere50", 
                        "MOMENT1024", "MOMENT50"))

ggplot(roc_data, aes(x = specificity, y = sensitivity, color = Model)) +
  geom_line() +
  theme( 
    legend.spacing.x = unit(2, "cm"),
    legend.justification = "left",   
    legend.box.just = "left",        
    legend.text = element_text(hjust = 0),
    panel.grid.major = element_blank(),  
    panel.grid.minor = element_blank()         
  ) +
  labs(x = "Specificity", y = "Sensitivity") 
```


## when importing got messed up 
```{r eval=FALSE}
# GPT with 1536 embedding dimension - from embedding_comparing.Rmd file 
data_gpt1536_raw = read.csv("./data2/data_wide_embedding_gpt1536.csv") |>
  janitor::clean_names() 
  #mutate(pair_id = rep(1:(n() / 2), each = 2))

demo_gpt = data_gpt1536_raw |>
  filter(row_number() %% 2 == 1) |>
  select(seqn, gender, age, race, education, married, pir, bmi, pair_id)

squished_rows = data_gpt1536_raw |>
  filter(row_number() %% 2 == 1) |>
  select(embedding, pair_id)

squished_list = squished_rows |>
  mutate(
    embedding_vec = map(embedding, function(x) {
      if (is.na(x)) return(NA)
      clean_str = str_remove_all(x, "\\[|\\]") |> 
        str_trim()
      
      if (nchar(clean_str) == 0) return(numeric(0))
      vals = str_split(clean_str, ",\\s*")[[1]]
      num_vals = suppressWarnings(as.numeric(vals))
      
      return(num_vals)
    })
  ) |>
  filter(!map_lgl(embedding_vec, function(v) length(v) == 1 && is.na(v))) |>
  select(-embedding)

extended_rows = data_gpt1536_raw |>
  filter(row_number() %% 2 == 0) |>
  select(-x) |>
  mutate(across(everything(), ~ str_remove(.x, "]\"$"))) 

colnames(extended_rows)[-which(colnames(extended_rows) == "pair_id")] = 
  paste0("x", seq_len(ncol(extended_rows) - 1))

extended_rows = extended_rows |>
  mutate(across(everything(), ~ as.numeric(.x))) 

extended_list = extended_rows |>
  group_by(pair_id) |>
  summarize(embedding_vec2 = list(as.numeric(unlist(across(where(is.numeric)))))
  )

demo_squish = demo_gpt |>
  left_join(squished_list, by = "pair_id")
  
data_gpt1536_long = demo_squish |>
  left_join(extended_list, by = "pair_id") |>
  mutate(c_embedding = map2(embedding_vec, embedding_vec2, c)) |>
  select(-embedding_vec, -embedding_vec2) |>
  unnest(c_embedding) |>
  group_by(seqn) |>
  mutate(var_id = row_number()) |>
  ungroup()

data_gpt1536 = data_gpt1536_long |>
  pivot_wider(
    names_from = var_id, 
    values_from = c_embedding, 
    names_prefix = "var"
  ) |>
  select(seqn, gender, age, race, education, married, pir, bmi, var1:var1536)

```


## embedding reduced to 50 dimension as EntroLLM method  
```{r}
# GPT with 50 embedding dimension
data_gpt50 = read.csv("./003_entro_data/data_wide_embedding_gpt50.csv") |>
  janitor::clean_names() |>
  dplyr::select(-combined,-n_tokens) 

# BERT with 50 embedding dimension
data_bert50 = read.csv("./003_entro_data/data_wide_embedding_bert50.csv") %>% 
  janitor::clean_names() |>
  dplyr::select(-x,-combined,-n_tokens) 

# Cohere with 50 embedding dimension
data_cohere50 = read.csv("./003_entro_data/data_wide_embedding_cohere50.csv") |>
  janitor::clean_names() |>
  dplyr::select(-x,-combined,-n_tokens)


# GPT50 + entropy
data_gpt50_entropy = data_entropy |>
  dplyr::select(seqn, entropy_day1:entropy_day7) |> 
  inner_join(data_gpt50, by = "seqn") |>
  left_join(covariates, by = "seqn") |>
  drop_na()

# BERT50 + entropy
data_bert50_entropy = data_entropy |>
  dplyr::select(seqn, entropy_day1:entropy_day7) |> 
  inner_join(data_bert50, by = "seqn") |>
  left_join(covariates, by = "seqn") |>
  drop_na()

# Cohere50 + entropy
data_cohere50_entropy = data_entropy |>
  dplyr::select(seqn, entropy_day1:entropy_day7) |> 
  inner_join(data_cohere50, by = "seqn") |>
  left_join(covariates, by = "seqn") |>
  drop_na()


# MOMENT with 50 embedding dimension 
data_moment1_50 = read_csv("./002_data/embeddings_moment_subset1_50.csv") |>
  janitor::clean_names() |>
  select(-x1) |>
  rename(x1 = x1_2) |>
  arrange(seqn) 

data_moment2_50 = read_csv("./002_data/embeddings_moment_subset2_50.csv") |>
  janitor::clean_names() |>
  select(-x1) |>
  rename(x1 = x1_2) |>
  arrange(seqn) 

data_moment50 = rbind(data_moment1_50, data_moment2_50)
```


## Visualization of PCA using demographics 
### 1. PCA of EntroGPT : using demographics 
```{r PCA_EntroGPT}
# change continuous age to categories 
data_gpt1536_entropy = data_gpt1536_entropy |>
    mutate(age_cat = case_when(age >= 20 & age <= 39 ~ 0,
                               age >= 40 & age <= 64 ~ 1,
                               age >= 65             ~ 2))

# separate into embeddings/entropy and covariates/response
entrogpt_sub1_embed = data_gpt1536_entropy[, c(2:8, 16:1551)]
entrogpt_sub2_var = data_gpt1536_entropy[, c(1,9:15, 1552)]

# scale the embeddings and combine with covariates 
entrogpt_sub1_embed = as.data.frame(scale(entrogpt_sub1_embed))
entrogpt_sub1_embed$seqn = data_gpt1536_entropy$seqn

entrogpt_scaled = entrogpt_sub1_embed |>
  left_join(entrogpt_sub2_var, by = "seqn") 


# PCA1 color by response (bmi) -- keep only embeddings and demo. 
pca_entrogpt_vars1 = entrogpt_scaled |>
  select(c(1:1543, 1545:1550)) #1544:seqn, 1551:bmi

pca_entrogpt1 = prcomp(pca_entrogpt_vars1, scale. = TRUE)

fviz_pca_ind(pca_entrogpt1, 
             habillage = ifelse(entrogpt_scaled$bmi==1, 
                                "BMI>25", "Otherwise"), 
             label = c("var"))

# PCA2 color by gender
pca_entrogpt_vars2 = entrogpt_scaled |>
  select(c(1:1543, 1546:1551)) #1545:gender

pca_entrogpt2 = prcomp(pca_entrogpt_vars2, scale. = TRUE)

fviz_pca_ind(pca_entrogpt2, 
             habillage = ifelse(entrogpt_scaled$gender==1, 
                                "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_entrogpt_vars3 = entrogpt_scaled |>
  select(c(1:1543, 1545, 1546, 1548:1551)) #1547:race

pca_entrogpt3 = prcomp(pca_entrogpt_vars3, scale. = TRUE)

fviz_pca_ind(pca_entrogpt3, 
             habillage = factor(entrogpt_scaled$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", 
                                         "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA4 color by education
pca_entrogpt_vars4 = entrogpt_scaled |>
  select(c(1:1543, 1545:1547, 1549:1551)) #1548:education

pca_entrogpt4 = prcomp(pca_entrogpt_vars4, scale. = TRUE)
fviz_pca_ind(pca_entrogpt4, 
             habillage = factor(entrogpt_scaled$education,
                               levels = 1:5,
                               labels = c("less than 9th grade", 
                                          "9-11th grade", 
                                         "high school grad/GED or equivalent", 
                                         "some college or associates degree,", 
                                         "college graduate or higher")), 
             label = c("var"))


# PCA5 color by married status
pca_entrogpt_vars5 = entrogpt_scaled |>
  select(c(1:1543, 1545:1548, 1550:1551)) #1549:married 

pca_entrogpt5 = prcomp(pca_entrogpt_vars5, scale. = TRUE)
fviz_pca_ind(pca_entrogpt5, 
             habillage = factor(entrogpt_scaled$married,
                               levels = 1:6,
                               labels = c("married", "widowed", 
                                         "divorced", "separated",
                                         "never married", "living with partner")),
             label = c("var"))


# PCA6 color by age
pca_entrogpt_vars6 = entrogpt_scaled |>
  select(c(1:1543, 1545, 1547:1551)) #1546:age 

pca_entrogpt6 = prcomp(pca_entrogpt_vars6, scale. = TRUE)
fviz_pca_ind(pca_entrogpt6, 
             habillage = factor(entrogpt_scaled$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", "older")),
             label = c("var"))
```


### 2. PCA of EntroBERT : using demographics 
```{r PCA_EntroBERT}
# change age to categories 
data_bert768_entropy = data_bert768_entropy |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# separate into embeddings/entropy and covariates/response
entrobert_sub1_embed = data_bert768_entropy[, c(2:8, 16:783)]
entrobert_sub2_var = data_bert768_entropy[, c(1, 9:15, 784)]

# scale the embeddings and combine with covariates 
entrobert_sub1_embed = as.data.frame(scale(entrobert_sub1_embed))
entrobert_sub1_embed$seqn = data_bert768_entropy$seqn

entrobert_scaled = entrobert_sub1_embed |>
  left_join(entrobert_sub2_var, by = "seqn") 



# PCA1 color by response (bmi)
pca_entrobert_vars1 = entrobert_scaled |>
  select(c(1:775, 777:782)) #776:seqn, 783:bmi

pca_entrobert1 = prcomp(pca_entrobert_vars1, scale. = TRUE)
fviz_pca_ind(pca_entrobert1, 
             habillage = ifelse(entrobert_scaled$bmi==1, 
                                "BMI>25", "Otherwise"), 
             label = c("var"))

# PCA2 color by gender
pca_entrobert_vars2 = entrobert_scaled |>
  select(c(1:775, 778:783)) #777:gender

pca_entrobert2 = prcomp(pca_entrobert_vars2, scale. = TRUE)
fviz_pca_ind(pca_entrobert2, 
             habillage = ifelse(entrobert_scaled$gender==1, "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_entrobert_vars3 = entrobert_scaled |>
  select(c(1:775, 777, 778, 780:783)) #779:race

pca_entrobert3 = prcomp(pca_entrobert_vars3, scale. = TRUE)
fviz_pca_ind(pca_entrobert3, 
             habillage = factor(entrobert_scaled$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA6 color by age
pca_entrobert_vars6 = entrobert_scaled |>
  select(c(1:775, 777, 779:783)) #778:age 

pca_entrobert6 = prcomp(pca_entrobert_vars6, scale. = TRUE)
fviz_pca_ind(pca_entrobert6, 
             habillage = factor(entrobert_scaled$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))
```



### 3. PCA of EntroCohere : using demographics 
```{r PCA_EntroCohere}
# change age to categories 
data_cohere1024_entropy = data_cohere1024_entropy |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# separate into embeddings/entropy and covariates/response
entrocohere_sub1_embed = data_cohere1024_entropy[, c(2:8, 16:1039)]
entrocohere_sub2_cov = data_cohere1024_entropy[, c(1,9:15, 1049)]

# scale the embeddings and combine with covariates 
entrocohere_sub1_embed = as.data.frame(scale(entrocohere_sub1_embed))
entrocohere_sub1_embed$seqn = data_cohere1024_entropy$seqn

entrocohere_scaled = entrocohere_sub1_embed |>
  left_join(entrocohere_sub2_cov, by = "seqn") 



# PCA1 color by response (bmi)
pca_entrocohere_vars1 = entrocohere_scaled |>
  select(c(1:1031, 1033:1038)) #1032:seqn, 1039:bmi

pca_entrocohere1 = prcomp(pca_entrocohere_vars1, scale. = TRUE)
fviz_pca_ind(pca_entrocohere1, 
             habillage = ifelse(entrocohere_scaled$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# PCA2 color by gender
pca_entrocohere_vars2 = entrocohere_scaled |>
  select(c(1:1031, 1034:1039)) #1033: gender   

pca_entrocohere2 = prcomp(pca_entrocohere_vars2, scale. = TRUE)
fviz_pca_ind(pca_entrocohere2, 
             habillage = ifelse(entrocohere_scaled$gender==1, "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_entrocohere_vars3 = entrocohere_scaled |>
  select(c(1:1031, 1033:1034, 1036:1039)) #1035: race 

pca_entrocohere3 = prcomp(pca_entrocohere_vars3, scale. = TRUE)
fviz_pca_ind(pca_entrocohere3, 
             habillage = factor(entrocohere_scaled$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA6 color by age
pca_entrocohere_vars6 = entrocohere_scaled |>
  select(c(1:1031, 1033, 1035:1039)) #1034:age  

pca_entrocohere6 = prcomp(entrocohere_scaled, scale. = TRUE)
fviz_pca_ind(pca_entrocohere6, 
             habillage = factor(entrocohere_combined$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))
```

### 4. PCA of MOMENT : using demographics 
```{r PCA_moment1024}
# change age to categories 
data_moment1024 = data_moment1024 |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# separate into embeddings/entropy and covariates/response
moment_sub1_embed = data_moment1024[, c(9:1032)]
moment_sub2_var = data_moment1024[, c(1:8, 1033:1042)]

# scale the embeddings and combine with covariates 
moment_sub1_embed = as.data.frame(scale(moment_sub1_embed))
moment_sub1_embed$seqn = data_moment1024$seqn
moment_scaled = moment_sub1_embed |>
  left_join(moment_sub2_var, by = "seqn")

# PCA1 color by response (bmi)
pca_moment_vars1 = moment_scaled |>
  select(c(1:1024, 1026:1031)) #1025:seqn, 1032:bmi 

pca_moment1 = prcomp(pca_moment_vars1, scale. = TRUE)
fviz_pca_ind(pca_moment1, 
             habillage = ifelse(moment_scaled$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# PCA2 color by gender
pca_moment_vars2 = moment_scaled |>
  select(c(1:1024, 1027:1032)) #1026:gender

pca_moment2 = prcomp(pca_moment_vars2, scale. = TRUE)
fviz_pca_ind(pca_moment2, 
             habillage = ifelse(moment_scaled$gender==1, "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_moment_vars3 = moment_scaled |>
  select(c(1:1024, 1026:1027, 1029:1032)) #1028:race 

pca_moment3 = prcomp(pca_moment_vars3, scale. = TRUE)
fviz_pca_ind(pca_moment3, 
             habillage = factor(moment_scaled$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA4 color by education
pca_moment_vars4 = moment_scaled |>
  select(c(1:1024, 1026:1028, 1030:1032)) #1029: education  

pca_moment4 = prcomp(pca_moment_vars4, scale. = TRUE)
fviz_pca_ind(pca_moment4, 
             habillage = factor(moment_scaled$education,
                               levels = 1:5,
                               labels = c("less than 9th grade", "9-11th grade", 
                                         "high school grad/GED or equivalent", 
                                         "some college or associates (AA) degree,", 
                                         "college graduate or higher")), 
             label = c("var"))

# PCA5 color by married status
pca_moment_vars5 = moment_scaled |>
  select(c(1:1024, 1026:1029, 1031:1032)) #1030:married 

pca_moment5 = prcomp(pca_moment_vars5, scale. = TRUE)
fviz_pca_ind(pca_moment5, 
             habillage = factor(moment_scaled$married,
                               levels = 1:6,
                               labels = c("married", "widowed", 
                                         "divorced", "separated",
                                         "never married", "living with partner")),
             label = c("var"))

# PCA6 color by age
pca_moment_vars6 = moment_scaled |>
  select(c(1:1024, 1026, 1028:1032)) #1027:age  

pca_moment6 = prcomp(pca_moment_vars6, scale. = TRUE)
fviz_pca_ind(pca_moment6, 
             habillage = factor(moment_scaled$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))
```

### 5. PCA of RAW DATA MOMENT embedding : using demographics 
```{r PCA_moment_raw1024}
# change age to categories 
data_raw_moment1024 = data_raw_moment1024 |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# separate into embeddings/entropy and covariates/response
moment_raw_embed = data_raw_moment1024[, c(9:1032)]
moment_raw_var = data_raw_moment1024[, c(1:8, 1033:1042)]

# scale the embeddings and combine with covariates 
moment_raw_embed = as.data.frame(scale(moment_raw_embed))
moment_raw_embed$seqn = data_raw_moment1024$seqn
moment_raw_scaled = moment_raw_embed |>
  left_join(moment_raw_var, by = "seqn")

# PCA1 color by response (bmi)
pca_moment_raw_vars1 = moment_raw_scaled |>
  select(c(1:1024, 1026:1031)) #1025:seqn, 1032:bmi 

pca_moment_raw1 = prcomp(pca_moment_raw_vars1, scale. = TRUE)
fviz_pca_ind(pca_moment_raw1, 
             habillage = ifelse(moment_raw_scaled$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# PCA2 color by gender
pca_moment_raw_vars2 = moment_raw_scaled |>
  select(c(1:1024, 1027:1032)) #1026:gender

pca_moment_raw2 = prcomp(pca_moment_raw_vars2, scale. = TRUE)
fviz_pca_ind(pca_moment_raw2, 
             habillage = ifelse(moment_raw_scaled$gender==1, "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_moment_raw_vars3 = moment_raw_scaled |>
  select(c(1:1024, 1026:1027, 1029:1032)) #1028:race 

pca_moment_raw3 = prcomp(pca_moment_raw_vars3, scale. = TRUE)
fviz_pca_ind(pca_moment_raw3, 
             habillage = factor(moment_raw_scaled$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA4 color by education
pca_moment_raw_vars4 = moment_raw_scaled |>
  select(c(1:1024, 1026:1028, 1030:1032)) #1029: education 

pca_moment_raw4 = prcomp(pca_moment_raw_vars4, scale. = TRUE)
fviz_pca_ind(pca_moment_raw4, 
             habillage = factor(moment_raw_scaled$education,
                               levels = 1:5,
                               labels = c("less than 9th grade", "9-11th grade", 
                                         "high school grad/GED or equivalent", 
                                         "some college or associates (AA) degree,", 
                                         "college graduate or higher")), 
             label = c("var"))

# PCA5 color by married status
pca_moment_raw_vars5 = moment_raw_scaled |>
  select(c(1:1024, 1026:1029, 1031:1032)) #1030:married 

pca_moment_raw5 = prcomp(pca_moment_raw_vars5, scale. = TRUE)
fviz_pca_ind(pca_moment_raw5, 
             habillage = factor(moment_raw_scaled$married,
                               levels = 1:6,
                               labels = c("married", "widowed", 
                                         "divorced", "separated",
                                         "never married", "living with partner")),
             label = c("var"))

# PCA6 color by age
pca_moment_raw_vars6 = moment_raw_scaled |>
  select(c(1:1024, 1026, 1028:1032)) #1027:age 

pca_moment_raw6 = prcomp(pca_moment_raw_vars6, scale. = TRUE)
fviz_pca_ind(pca_moment_raw6, 
             habillage = factor(moment_raw_scaled$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))
```

### 6. PCA of RAW RECODED MOMENT embedding : using demographics 
```{r PCA_moment_recoded1024}
# add more covariates 
data_moment_recoded1024 = data_moment_recoded1024 |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# separate into embeddings/entropy and covariates/response
moment_recoded_embed = data_moment_recoded1024[, c(9:1032)]
moment_recoded_cov = data_moment_recoded1024[, c(1:8, 1033)]

# scale the embeddings and combine with covariates 
moment_recoded_embed = as.data.frame(scale(moment_recoded_embed))
moment_recoded_embed$seqn = data_moment_recoded1024$seqn
moment_recoded_combined = moment_recoded_embed |>
  left_join(moment_recoded_cov, by = "seqn")

# PCA1 color by response (bmi)
pca_moment_recoded_vars1 = moment_recoded_combined |>
  select(-seqn, -age_cat, -bmi)  
pca_moment_recoded1 = prcomp(pca_moment_recoded_vars1, scale. = TRUE)
fviz_pca_ind(pca_moment_recoded1, 
             habillage = ifelse(moment_recoded_combined$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# PCA2 color by gender
pca_moment_recoded_vars2 = moment_recoded_combined |>
  select(-seqn, -age_cat, -gender)  
pca_moment_recoded2 = prcomp(pca_moment_recoded_vars2, scale. = TRUE)
fviz_pca_ind(pca_moment_recoded2, 
             habillage = ifelse(moment_recoded_combined$gender==1, "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_moment_recoded_vars3 = moment_recoded_combined |>
  select(-seqn, -age_cat, -race)  
pca_moment_recoded3 = prcomp(pca_moment_recoded_vars3, scale. = TRUE)
fviz_pca_ind(pca_moment_recoded3, 
             habillage = factor(moment_recoded_combined$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA4 color by education
pca_moment_recoded_vars4 = moment_recoded_combined |>
  select(-seqn, -age_cat, -education)  
pca_moment_recoded4 = prcomp(pca_moment_recoded_vars4, scale. = TRUE)
fviz_pca_ind(pca_moment_recoded4, 
             habillage = factor(moment_recoded_combined$education,
                               levels = 1:5,
                               labels = c("less than 9th grade", "9-11th grade", 
                                         "high school grad/GED or equivalent", 
                                         "some college or associates (AA) degree,", 
                                         "college graduate or higher")), 
             label = c("var"))

# PCA5 color by married status
pca_moment_recoded_vars5 = moment_recoded_combined |>
  select(-seqn, -age_cat, -married)  
pca_moment_recoded5 = prcomp(pca_moment_recoded_vars5, scale. = TRUE)
fviz_pca_ind(pca_moment_recoded5, 
             habillage = factor(moment_recoded_combined$married,
                               levels = 1:6,
                               labels = c("married", "widowed", 
                                         "divorced", "separated",
                                         "never married", "living with partner")),
             label = c("var"))

# PCA6 color by age
pca_moment_recoded_vars6 = moment_recoded_combined |>
  select(-seqn, -age, -age_cat)
pca_moment_recoded6 = prcomp(pca_moment_recoded_vars6, scale. = TRUE)
fviz_pca_ind(pca_moment_recoded6, 
             habillage = factor(moment_recoded_combined$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))
```



### 7. PCA of the 5 min interval data by bmi, gender, race, etc. 
```{r PCA_original, eval=FALSE}
# import original 5 minute interval data and group by age
# young (20–39), middle (40–64), older (65+) 
original_5min = read.csv("./data/data_wide.csv") |>
  janitor::clean_names() |>
  select(-x) |>
  mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# separate into PA/entropy and covariates/response
original_5min_pa = original_5min[, c(9:2024)]
original_5min_cov = original_5min[, c(1:8, 2025)]

# scale the PA and combine with covariates 
original_5min_pa = as.data.frame(scale(original_5min_pa))
original_5min_pa$seqn = original_5min$seqn
original_5min_combined = original_5min_pa |>
  left_join(original_5min_cov, by = "seqn")

# PCA1 color by response (bmi)
pca_5min_vars1 = original_5min_combined |>
  select(-seqn, -age_cat, -bmi)
pca_5min1 = prcomp(pca_5min_vars1, scale. = TRUE)
fviz_pca_ind(pca_5min1,
             geom.ind = "point",
             habillage = ifelse(original_5min_combined$bmi==1, "BMI>25", "Otherwise"), 
             label = "var")

# PCA2 color by gender
pca_5min_vars2 = original_5min_combined |>
  select(-seqn, -age_cat, -gender)  
pca_5min2 = prcomp(pca_5min_vars2, scale. = TRUE)
fviz_pca_ind(pca_5min2, 
             habillage = ifelse(original_5min_combined$gender==1, "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_5min_vars3 = original_5min_combined |>
  select(-seqn, -age_cat, -race)  
pca_5min3 = prcomp(pca_5min_vars3, scale. = TRUE)
fviz_pca_ind(pca_5min3, 
             habillage = factor(original_5min_combined$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA4 color by education
pca_5min_vars4 = original_5min_combined |>
  select(-seqn, -age_cat, -education)  
pca_5min4 = prcomp(pca_5min_vars4, scale. = TRUE)
fviz_pca_ind(pca_5min4, 
             habillage = factor(original_5min_combined$education,
                               levels = 1:5,
                               labels = c("less than 9th grade", "9-11th grade", 
                                         "high school grad/GED or equivalent", 
                                         "some college or associates (AA) degree,", 
                                         "college graduate or higher")), 
             label = c("var"))

# PCA5 color by married status
pca_5min_vars5 = original_5min_combined |>
  select(-seqn, -age_cat, -married)  
pca_5min5 = prcomp(pca_5min_vars5, scale. = TRUE)
fviz_pca_ind(pca_5min5, 
             habillage = factor(original_5min_combined$married,
                               levels = 1:6,
                               labels = c("married", "widowed", 
                                         "divorced", "separated",
                                         "never married", "living with partner")),
             label = c("var"))

# PCA6 color by age
pca_5min_vars6 = original_5min_combined |>
  select(-seqn, -age, -age_cat)
pca_5min6 = prcomp(pca_5min_vars6, scale. = TRUE)
fviz_pca_ind(pca_5min6, 
             habillage = factor(original_5min_combined$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))
```


### 9. Check PCA with 20 dimensions 
```{r eval=FALSE} 
# EntroGPT
pca20_entrogpt = prcomp(pca_entrogpt_vars1, scale. = TRUE, rank = 20)
fviz_pca_ind(pca20_entrogpt, 
             habillage = ifelse(entrogpt_combined$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# EntroBERT
pca20_entrobert = prcomp(pca_entrobert_vars1, scale. = TRUE, rank = 20)
fviz_pca_ind(pca20_entrobert, 
             habillage = ifelse(entrogpt_combined$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# EntroCOHERE
pca20_entrocohere = prcomp(pca_entrochere_vars1, scale. = TRUE, rank = 20)
fviz_pca_ind(pca20_entrocohere, 
             habillage = ifelse(entrocohere_combined$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# MOMENT
pca20_moment = prcomp(pca_moment_vars1, scale. = TRUE, rank = 20)
fviz_pca_ind(pca20_moment, 
             habillage = ifelse(moment_combined$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

```



## Visualization - PCA with all covaraites and scaled emebddings  

### 1. PCA of EntroGPT  
```{r PCA_EntroGPT with}
# PCA with embeddings/all covaraites of 2914 data points and color by bmi 
pca_entrogpt_vars_covall1 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1550, 1553:1561)) #1544:seqn, 1551:bmi, 1552:age_cat, 1562:1566 converted categories 

pca_entrogpt_covall1 = prcomp(pca_entrogpt_vars_covall1, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall1, 
             habillage = ifelse(entrogpt_scaled_cov$bmi==1, 
                                "BMI>25", "Otherwise"), 
             label = c("var"))


# PCA with embeddings/all covaraites of 2914 data points and color by gender 
pca_entrogpt_vars_covall2 = entrogpt_scaled_cov |>
  select(c(1:1543, 1546:1551, 1553:1561)) #1544:seqn, 1545:gender, 1552:age_cat, 1562:1566 converted categories 

pca_entrogpt_covall2 = prcomp(pca_entrogpt_vars_covall2, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall2, 
             habillage = ifelse(entrogpt_scaled_cov$gender==1, "male", "female"), 
             label = c("var"))


# PCA with embeddings/all covaraites of 2914 data points and color by race
pca_entrogpt_vars_covall3 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1546, 1548:1551, 1553:1561)) #1544:seqn, 1547:race, 1552:age_cat, 1562:1566 converted categories 

pca_entrogpt_covall3 = prcomp(pca_entrogpt_vars_covall3, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall3, 
             habillage = factor(entrogpt_scaled_cov$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", 
                                         "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))



# PCA6 color by age
pca_entrogpt_vars_covall6 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545, 1547:1551, 1553:1561)) #1544:seqn, 1546:age, 1552:age_cat, 1562:1566 converted categories 

pca_entrogpt_covall6 = prcomp(pca_entrogpt_vars_covall6, scale. = TRUE)
fviz_pca_ind(pca_entrogpt_covall6, 
             habillage = factor(entrogpt_scaled_cov$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))


# color by systolic BP
pca_entrogpt_vars_covall7 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1551, 1554:1561)) #1544:seqn, 1552:age_cat, 1553:med_sbp,  1562:1566 converted categories 
pca_entrogpt_covall7 = prcomp(pca_entrogpt_vars_covall7, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall7, 
             habillage = factor(entrogpt_scaled_cov$sbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by diastolic BP
pca_entrogpt_vars_covall8 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553, 1555:1561)) #1544:seqn, 1552:age_cat, 1554:med_dbp,  1562:1566 converted categories 
pca_entrogpt_covall8 = prcomp(pca_entrogpt_vars_covall8, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall8, 
             habillage = factor(entrogpt_scaled_cov$dbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by total cholesterol
pca_entrogpt_vars_covall9 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1554, 1556:1561)) #1544:seqn, 1552:age_cat, 1555:chol_total_cat,  1562:1566 converted categories 
pca_entrogpt_covall9 = prcomp(pca_entrogpt_vars_covall9, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall9, 
             habillage = factor(entrogpt_scaled_cov$chol_total_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by HDL 
pca_entrogpt_vars_covall10 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1555, 1557:1561)) #1544:seqn, 1552:age_cat, 1556:chol_hdl_cat,  1562:1566 converted categories 
pca_entrogpt_covall10 = prcomp(pca_entrogpt_vars_covall10, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall10, 
             habillage = factor(entrogpt_scaled_cov$chol_hdl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by LDL 
pca_entrogpt_vars_covall11 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1556, 1558:1561)) #1544:seqn, 1552:age_cat, 1557:chol_ldl_cat,  1562:1566 converted categories 
pca_entrogpt_covall11 = prcomp(pca_entrogpt_vars_covall11, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall11, 
             habillage = factor(entrogpt_scaled_cov$chol_ldl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by diabetes  
pca_entrogpt_vars_covall12 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1557, 1559:1561)) #1544:seqn, 1552:age_cat, 1558:diabetes,  1562:1566 converted categories 
pca_entrogpt_covall12 = prcomp(pca_entrogpt_vars_covall12, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall12, 
             habillage = factor(entrogpt_scaled_cov$diabetes,
                               levels = 1:3,
                               labels = c("yes", "no", 
                                           "borderline")),  
             label = c("var"))

# color by arthritis  
pca_entrogpt_vars_covall13 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1558, 1560:1561)) #1544:seqn, 1552:age_cat, 1559:arthritis,  1562:1566 converted categories 
pca_entrogpt_covall13 = prcomp(pca_entrogpt_vars_covall13, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall13, 
             habillage = ifelse(entrogpt_scaled_cov$arthritis==1, 
                                "Arthritis", "No"), 
             label = c("var"))

# color by overweight 
pca_entrogpt_vars_covall14 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1559, 1561)) #1544:seqn, 1552:age_cat, 1560:overweight,  1562:1566 converted categories 
pca_entrogpt_covall14 = prcomp(pca_entrogpt_vars_covall14, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_vars_covall14, 
             habillage = ifelse(entrogpt_scaled_cov$overweight==1, 
                                "Overweight", "No"), 
             label = c("var"))

# color by malignancy  
pca_entrogpt_vars_covall15 = entrogpt_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1560)) #1544:seqn, 1552:age_cat, 1561:overweight,  1562:1566 converted categories 
pca_entrogpt_covall15 = prcomp(pca_entrogpt_vars_covall15, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall15, 
             habillage = ifelse(entrogpt_scaled_cov$malig==1, 
                                "Malignancy", "No"), 
             label = c("var"))
```

### 2. PCA of EntroBert  
```{r PCA_EntroBERT_with}
# PCA with embeddings/all covaraites of 2914 data points and color by bmi 
pca_entrobert_vars_covall1 = entrobert_scaled_cov |>
  select(c(1:775, 777:782, 785:793)) #776:seqn, 783:bmi, 784:age_cat, 794:798 categorized 

pca_entrobert_covall1 = prcomp(pca_entrobert_vars_covall1, scale. = TRUE)
fviz_pca_ind(pca_entrobert_covall1, 
             habillage = ifelse(entrobert_scaled_cov$bmi==1, 
                                "BMI>25", "Otherwise"), 
             label = c("var"))


# PCA2 color by gender
pca_entrobert_vars_covall2 = entrobert_scaled_cov |>
  select(c(1:775, 778:783, 785:793)) #776:seqn, 777:gender, 784:age_cat, 794:798 categorized 

pca_entrobert_covall2 = prcomp(pca_entrobert_vars_covall2, scale. = TRUE)
fviz_pca_ind(pca_entrobert_covall2, 
             habillage = ifelse(entrobert_scaled_cov$gender==1, 
                                "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_entrobert_vars_covall3 = entrobert_scaled_cov |>
  select(c(1:775, 777:778, 780:783, 785:793)) #776:seqn, 779:race, 784:age_cat, 794:798 categorized 

pca_entrobert_covall3 = prcomp(pca_entrobert_vars_covall3, scale. = TRUE)
fviz_pca_ind(pca_entrobert_covall3, 
             habillage = factor(entrobert_scaled_cov$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA6 color by age
pca_entrobert_vars_covall6 = entrobert_scaled_cov |>
  select(c(1:775, 777, 779:783, 785:793)) #776:seqn, 778:age, 784:age_cat, 794:798 categorized 

pca_entrobert_covall6 = prcomp(pca_entrobert_vars_covall6, scale. = TRUE)
fviz_pca_ind(pca_entrobert_covall6, 
             habillage = factor(entrobert_scaled_cov$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))




# color by systolic BP
pca_entrobert_vars_covall7 = entrobert_scaled_cov |>
  select(c(1:775, 777:783, 785:793)) #776:seqn, 785:med_sbp, 784:age_cat, 794:798 categorized 
pca_entrogpt_covall7 = prcomp(pca_entrobert_vars_covall7, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall7, 
             habillage = factor(entrobert_scaled_cov$sbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))


############ STOP HERE 
# color by diastolic BP
pca_entrobert_vars_covall8 = entrobert_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553, 1555:1561)) #1544:seqn, 1552:age_cat, 1554:med_dbp,  1562:1566 converted categories 
pca_entrogpt_covall8 = prcomp(pca_entrobert_vars_covall8, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall8, 
             habillage = factor(entrobert_scaled_cov$dbp_cat,
                               levels = 0:3,
                               labels = c("low risk", "stage1 hypertension", 
                                         "stage2 hypertension", "crisis")),
             label = c("var"))

# color by total cholesterol
pca_entrogpt_vars_covall9 = entrobert_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1554, 1556:1561)) #1544:seqn, 1552:age_cat, 1555:chol_total_cat,  1562:1566 converted categories 
pca_entrogpt_covall9 = prcomp(pca_entrogpt_vars_covall9, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall9, 
             habillage = factor(entrobert_scaled_cov$chol_total_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by HDL 
pca_entrogpt_vars_covall10 = entrobert_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1555, 1557:1561)) #1544:seqn, 1552:age_cat, 1556:chol_hdl_cat,  1562:1566 converted categories 
pca_entrogpt_covall10 = prcomp(pca_entrogpt_vars_covall10, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall10, 
             habillage = factor(entrobert_scaled_cov$chol_hdl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by LDL 
pca_entrogpt_vars_covall11 = entrobert_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1556, 1558:1561)) #1544:seqn, 1552:age_cat, 1557:chol_ldl_cat,  1562:1566 converted categories 
pca_entrogpt_covall11 = prcomp(pca_entrogpt_vars_covall11, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall11, 
             habillage = factor(entrobert_scaled_cov$chol_ldl_cat,
                               levels = 0:2,
                               labels = c("low risk", "borderline", 
                                           "high risk")),  
             label = c("var"))

# color by diabetes  
pca_entrogpt_vars_covall12 = entrobert_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1557, 1559:1561)) #1544:seqn, 1552:age_cat, 1558:diabetes,  1562:1566 converted categories 
pca_entrogpt_covall12 = prcomp(pca_entrogpt_vars_covall12, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall12, 
             habillage = factor(entrobert_scaled_cov$diabetes,
                               levels = 1:3,
                               labels = c("yes", "no", 
                                           "borderline")),  
             label = c("var"))

# color by arthritis  
pca_entrogpt_vars_covall13 = entrobert_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1558, 1560:1561)) #1544:seqn, 1552:age_cat, 1559:arthritis,  1562:1566 converted categories 
pca_entrogpt_covall13 = prcomp(pca_entrogpt_vars_covall13, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall13, 
             habillage = ifelse(entrobert_scaled_cov$arthritis==1, 
                                "Arthritis", "No"), 
             label = c("var"))

# color by overweight 
pca_entrogpt_vars_covall14 = entrobert_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1559, 1561)) #1544:seqn, 1552:age_cat, 1560:overweight,  1562:1566 converted categories 
pca_entrogpt_covall14 = prcomp(pca_entrogpt_vars_covall14, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_vars_covall14, 
             habillage = ifelse(entrobert_scaled_cov$overweight==1, 
                                "Overweight", "No"), 
             label = c("var"))

# color by malignancy  
pca_entrogpt_vars_covall15 = entrobert_scaled_cov |>
  select(c(1:1543, 1545:1551, 1553:1560)) #1544:seqn, 1552:age_cat, 1561:overweight,  1562:1566 converted categories 
pca_entrogpt_covall15 = prcomp(pca_entrogpt_vars_covall15, scale. = TRUE)

fviz_pca_ind(pca_entrogpt_covall15, 
             habillage = ifelse(entrobert_scaled_cov$malig==1, 
                                "Malignancy", "No"), 
             label = c("var"))
```

### 3. PCA of EntroCohere  
```{r PCA_EntroCohere_with}
# change age to categories 
data_cohere1024_entropy = data_cohere1024_entropy |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# separate into embeddings/entropy and covariates/response
entrocohere_embed = data_cohere1024_entropy[, c(2:8, 16:1039)]
entrocohere_cov = data_cohere1024_entropy[, c(1,9:15, 1049)]

# scale the embeddings and combine with covariates 
entrocohere_embed = as.data.frame(scale(entrocohere_embed))
entrocohere_embed$seqn = data_cohere1024_entropy$seqn
entrocohere_combined = entrocohere_embed |>
  left_join(entrocohere_cov, by = "seqn") 

# PCA1 color by response (bmi)
pca_entrocohere_vars1 = entrocohere_combined |>
  select(c(1:1031, 1033:1038)) #1032:seqn, 1039:bmi

pca_entrocohere1 = prcomp(pca_entrocohere_vars1, scale. = TRUE)
fviz_pca_ind(pca_entrocohere1, 
             habillage = ifelse(entrocohere_combined$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# PCA2 color by gender
pca_entrocohere_vars2 = entrocohere_combined |>
  select(c(1:1031, 1034:1039)) #1033: gender   

pca_entrocohere2 = prcomp(pca_entrocohere_vars2, scale. = TRUE)
fviz_pca_ind(pca_entrocohere2, 
             habillage = ifelse(entrocohere_combined$gender==1, "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_entrocohere_vars3 = entrocohere_combined |>
  select(c(1:1031, 1033:1034, 1036:1039)) #1035: race 

pca_entrocohere3 = prcomp(pca_entrocohere_vars3, scale. = TRUE)
fviz_pca_ind(pca_entrocohere3, 
             habillage = factor(entrocohere_combined$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA6 color by age
pca_entrocohere_vars6 = entrocohere_combined |>
  select(c(1:1031, 1033, 1035:1039)) #1034:age  

pca_entrocohere6 = prcomp(pca_entrocohere_vars6, scale. = TRUE)
fviz_pca_ind(pca_entrocohere6, 
             habillage = factor(entrocohere_combined$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))
```

### 4. PCA of MOMENT 
```{r PCA_moment1024_with}
# change age to categories 
data_moment1024 = data_moment1024 |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# separate into embeddings/entropy and covariates/response
moment_embed = data_moment1024[, c(9:1032)]
moment_cov = data_moment1024[, c(1:8, 1033:1042)]

# scale the embeddings and combine with covariates 
moment_embed = as.data.frame(scale(moment_embed))
moment_embed$seqn = data_moment1024$seqn
moment_combined = moment_embed |>
  left_join(moment_cov, by = "seqn")

# PCA1 color by response (bmi)
pca_moment_vars1 = moment_combined |>
  select(c(1:1024, 1026:1031)) #1025:seqn, 1032:bmi 

pca_moment1 = prcomp(pca_moment_vars1, scale. = TRUE)
fviz_pca_ind(pca_moment1, 
             habillage = ifelse(moment_combined$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# PCA2 color by gender
pca_moment_vars2 = moment_combined |>
  select(c(1:1024, 1027:1032)) #1026:gender

pca_moment2 = prcomp(pca_moment_vars2, scale. = TRUE)
fviz_pca_ind(pca_moment2, 
             habillage = ifelse(moment_combined$gender==1, "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_moment_vars3 = moment_combined |>
  select(c(1:1024, 1026:1027, 1029:1032)) #1028:race 

pca_moment3 = prcomp(pca_moment_vars3, scale. = TRUE)
fviz_pca_ind(pca_moment3, 
             habillage = factor(moment_combined$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA4 color by education
pca_moment_vars4 = moment_combined |>
  select(c(1:1024, 1026:1028, 1030:1032)) #1029: education  

pca_moment4 = prcomp(pca_moment_vars4, scale. = TRUE)
fviz_pca_ind(pca_moment4, 
             habillage = factor(moment_combined$education,
                               levels = 1:5,
                               labels = c("less than 9th grade", "9-11th grade", 
                                         "high school grad/GED or equivalent", 
                                         "some college or associates (AA) degree,", 
                                         "college graduate or higher")), 
             label = c("var"))

# PCA5 color by married status
pca_moment_vars5 = moment_combined |>
  select(c(1:1024, 1026:1029, 1031:1032)) #1030:married 

pca_moment5 = prcomp(pca_moment_vars5, scale. = TRUE)
fviz_pca_ind(pca_moment5, 
             habillage = factor(moment_combined$married,
                               levels = 1:6,
                               labels = c("married", "widowed", 
                                         "divorced", "separated",
                                         "never married", "living with partner")),
             label = c("var"))

# PCA6 color by age
pca_moment_vars6 = moment_combined |>
  select(c(1:1024, 1026, 1028:1032)) #1027:age  

pca_moment6 = prcomp(pca_moment_vars6, scale. = TRUE)
fviz_pca_ind(pca_moment6, 
             habillage = factor(moment_combined$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))
```

### 5. PCA of MOMENT RAW embedding  
```{r PCA_moment_raw1024_with}
# change age to categories 
data_moment_raw1024 = data_moment_raw1024 |>
    mutate(age_cat = case_when(
      age >= 20 & age <= 39 ~ 0,
      age >= 40 & age <= 64 ~ 1,
      age >= 65             ~ 2))

# separate into embeddings/entropy and covariates/response
moment_raw_embed = data_moment_raw1024[, c(9:1032)]
moment_raw_cov = data_moment_raw1024[, c(1:8, 1033:1042)]

# scale the embeddings and combine with covariates 
moment_raw_embed = as.data.frame(scale(moment_raw_embed))
moment_raw_embed$seqn = data_moment_raw1024$seqn
moment_raw_combined = moment_raw_embed |>
  left_join(moment_raw_cov, by = "seqn")

# PCA1 color by response (bmi)
pca_moment_raw_vars1 = moment_raw_combined |>
  select(c(1:1024, 1026:1031)) #1025:seqn, 1032:bmi 

pca_moment_raw1 = prcomp(pca_moment_raw_vars1, scale. = TRUE)
fviz_pca_ind(pca_moment_raw1, 
             habillage = ifelse(moment_raw_combined$bmi==1, "BMI>25", "Otherwise"), 
             label = c("var"))

# PCA2 color by gender
pca_moment_raw_vars2 = moment_raw_combined |>
  select(c(1:1024, 1027:1032)) #1026:gender

pca_moment_raw2 = prcomp(pca_moment_raw_vars2, scale. = TRUE)
fviz_pca_ind(pca_moment_raw2, 
             habillage = ifelse(moment_raw_combined$gender==1, "male", "female"), 
             label = c("var"))

# PCA3 color by race
pca_moment_raw_vars3 = moment_raw_combined |>
  select(c(1:1024, 1026:1027, 1029:1032)) #1028:race 

pca_moment_raw3 = prcomp(pca_moment_raw_vars3, scale. = TRUE)
fviz_pca_ind(pca_moment_raw3, 
             habillage = factor(moment_raw_combined$race,
                               levels = 1:5,
                               labels = c("Mexican American", "Other Hispanic", 
                                         "Non-Hispanic White", "Non-Hispanic Black", 
                                         "Other")), 
             label = c("var"))

# PCA4 color by education
pca_moment_raw_vars4 = moment_raw_combined |>
  select(c(1:1024, 1026:1028, 1030:1032)) #1029: education 

pca_moment_raw4 = prcomp(pca_moment_raw_vars4, scale. = TRUE)
fviz_pca_ind(pca_moment_raw4, 
             habillage = factor(moment_raw_combined$education,
                               levels = 1:5,
                               labels = c("less than 9th grade", "9-11th grade", 
                                         "high school grad/GED or equivalent", 
                                         "some college or associates (AA) degree,", 
                                         "college graduate or higher")), 
             label = c("var"))

# PCA5 color by married status
pca_moment_raw_vars5 = moment_raw_combined |>
  select(c(1:1024, 1026:1029, 1031:1032)) #1030:married 

pca_moment_raw5 = prcomp(pca_moment_raw_vars5, scale. = TRUE)
fviz_pca_ind(pca_moment_raw5, 
             habillage = factor(moment_raw_combined$married,
                               levels = 1:6,
                               labels = c("married", "widowed", 
                                         "divorced", "separated",
                                         "never married", "living with partner")),
             label = c("var"))

# PCA6 color by age
pca_moment_raw_vars6 = moment_raw_combined |>
  select(c(1:1024, 1026, 1028:1032)) #1027:age 

pca_moment_raw6 = prcomp(pca_moment_raw_vars6, scale. = TRUE)
fviz_pca_ind(pca_moment_raw6, 
             habillage = factor(moment_raw_combined$age_cat,
                               levels = 0:2,
                               labels = c("young", "middle", 
                                         "older")),
             label = c("var"))
```
